Evaluate an expression in Civis Platform

Evaluate an expression in Civis Platform

CivisFuture(
  expr = NULL,
  envir = parent.frame(),
  substitute = FALSE,
  globals = TRUE,
  packages = NULL,
  lazy = FALSE,
  local = lifecycle::deprecated(),
  gc = FALSE,
  label = NULL,
  required_resources = list(cpu = 1024, memory = 2048, diskSpace = 4),
  docker_image_name = "civisanalytics/datascience-r",
  docker_image_tag = "latest",
  ...
)

# S3 method for class 'CivisFuture'
run(future, ...)

# S3 method for class 'CivisFuture'
result(future, ...)

# S3 method for class 'CivisFuture'
cancel(future, ...)

# S3 method for class 'CivisFuture'
resolved(future, ...)

# S3 method for class 'CivisFuture'
fetch_logs(object, ...)

Arguments

expr

An R expression.

envir

The environment from where global objects should be identified.

substitute

If TRUE, argument expr is substitute():ed, otherwise not.

globals

(optional) a logical, a character vector, or a named list to control how globals are handled. For details, see section 'Globals used by future expressions' in the help for future().

packages

(optional) a character vector specifying packages to be attached in the R environment evaluating the future.

lazy

If FALSE (default), the future is resolved eagerly (starting immediately), otherwise not.

local

deprecated as of civis v3.0.1

gc

If TRUE, the garbage collector is run after the value of the future has been collected, i.e., when it is no longer needed or when the future is terminated.

label

An optional character string label attached to the future.

required_resources

resources, see scripts_post_containers

docker_image_name

the image for the container script.

docker_image_tag

the tag for the Docker image.

...

arguments to scripts_post_containers

future

CivisFuture object.

object

CivisFuture

Value

A CivisFuture inheriting from Future that evaluates expr on the given container.

Methods (by generic)

run(CivisFuture): Run a CivisFuture

result(CivisFuture): Return the value of a CivisFuture

cancel(CivisFuture): Cancel a CivisFuture

resolved(CivisFuture): Check if a CivisFuture has resolved

fetch_logs(CivisFuture): Fetch logs from a CivisFuture