DBI methods

Implementations of pure virtual functions defined in the `DBI` package. The DBI connectors for Civis Platform are no longer maintained.

# S4 method for CivisDriver
show(object)

# S4 method for CivisDriver
dbConnect(drv, database, ...)

# S4 method for CivisDriver,ANY
dbDataType(dbObj, obj, ...)

# S4 method for CivisDriver,list
dbDataType(dbObj, obj, ...)

# S4 method for CivisDriver
dbIsValid(dbObj, ...)

# S4 method for CivisDriver
dbGetInfo(dbObj, ...)

# S4 method for CivisConnection
show(object)

# S4 method for CivisConnection
dbIsValid(dbObj, ...)

# S4 method for CivisConnection
dbIsReadOnly(dbObj, ...)

# S4 method for CivisConnection
dbDisconnect(conn, ...)

# S4 method for CivisConnection,character
dbSendQuery(conn, statement, ...)

# S4 method for CivisConnection,character
dbSendStatement(conn, statement, ...)

# S4 method for CivisConnection,ANY
dbDataType(dbObj, obj, ...)

# S4 method for CivisConnection,character
dbQuoteString(conn, x, ...)

# S4 method for CivisConnection,character
dbQuoteIdentifier(conn, x, ...)

# S4 method for CivisConnection,character,data.frame
dbWriteTable(conn, name, value,
  overwrite = FALSE, append = FALSE, ...)

# S4 method for CivisConnection,character
dbReadTable(conn, name, ...)

# S4 method for CivisConnection
dbListTables(conn, ...)

# S4 method for CivisConnection,character
dbExistsTable(conn, name, ...)

# S4 method for CivisConnection,character
dbListFields(conn, name, ...)

# S4 method for CivisConnection,character
dbRemoveTable(conn, name, ...)

# S4 method for CivisConnection
dbGetInfo(dbObj, ...)

# S4 method for CivisConnection
dbBegin(conn, ...)

# S4 method for CivisConnection
dbCommit(conn, ...)

# S4 method for CivisConnection
dbRollback(conn, ...)

# S4 method for CivisResult
show(object)

# S4 method for CivisResult
dbClearResult(res, ...)

# S4 method for CivisResult
dbFetch(res, n = -1, ...)

# S4 method for CivisResult
dbHasCompleted(res, ...)

# S4 method for CivisResult
dbGetInfo(dbObj, ...)

# S4 method for CivisResult
dbIsValid(dbObj, ...)

# S4 method for CivisResult
dbGetStatement(res, ...)

# S4 method for CivisResult
dbColumnInfo(res, ...)

# S4 method for CivisResult
dbGetRowCount(res, ...)

# S4 method for CivisResult
dbGetRowsAffected(res, ...)

# S4 method for CivisResult
dbBind(res, params, ...)

Arguments

object
Any R object

drv
an object that inherits from DBIDriver, or an existing DBIConnection object (in order to clone an existing connection).

database
The name of a database on the Civis Platform.

...
authentication arguments needed by the DBMS instance; these typically include user, password, host, port, dbname, etc. For details see the appropriate DBIDriver.

dbObj
A object inheriting from DBIDriver or DBIConnection

obj
An R object whose SQL type we want to determine.

conn
A DBIConnection object, as returned by dbConnect().

statement
a character string containing SQL.

x
A character vector to quote as string.

name
A character string specifying the unquoted DBMS table name, or the result of a call to dbQuoteIdentifier().

value
a data.frame (or coercible to data.frame).

overwrite
Allow overwriting the destination table. Cannot be `TRUE` if `append` is also `TRUE`.

append
Allow appending to the destination table. Cannot be `TRUE` if `overwrite` is also `TRUE`.

res
An object inheriting from DBIResult.

n
maximum number of records to retrieve per fetch. Use n = -1 or n = Inf to retrieve all pending records. Some implementations may recognize other special values.

params
A list of bindings, named or unnamed.