List organizations

List organizations

admin_list_organizations(status = NULL, org_type = NULL)

Arguments

status
array optional. The status of the organization (active/trial/inactive).

org_type
array optional. The organization type (platform/ads/survey_vendor/other).

Value

An array containing the following fields:

id

integer, The ID of this organization.

name

string, The name of this organization.

slug

string, The slug of this organization.

accountManagerId

integer, The user ID of the Account Manager.

csSpecialistId

integer, The user ID of the Client Success Specialist.

status

string, The status of the organization (active/trial/inactive).

orgType

string, The organization type (platform/ads/survey_vendor/other).

customBranding

string, The custom branding settings.

contractSize

integer, The monthly contract size.

maxAnalystUsers

integer, The max number of full platform users for the org.

maxReportUsers

integer, The max number of report-only platform users for the org.

vertical

string, The business vertical that the organization belongs to.

csMetadata

string, Additional metadata about the organization in JSON format.

removeFooterInEmails

boolean, If true, emails sent by platform will not include Civis text.

salesforceAccountId

string, The SalesForce Account ID for this organization.

advancedSettings

list, A list containing the following elements:

dedicatedDjPoolEnabled boolean, If true, the Organization has a dedicated delayed jobs pool. Defaults to false.

tableauRefreshHistory

array, The number of tableau refreshes used this month.