Create Bocce Cluster

Create Bocce Cluster

bocce_clusters_post(instance_type, min_instances, max_instances, region,
  single_org, organization_ids = NULL, desired_instances = NULL,
  asg_name = NULL, queue_url = NULL)

Arguments

instance_type
string required. The EC2 instance types in this cluster.

min_instances
integer required. The minimum number of instances in this cluster.

max_instances
integer required. The maximum number of instances in this cluster.

region
string required. The AWS region that this cluster is in.

single_org
boolean required. Whether this cluster is allowed to be used by multiple orgs.

organization_ids
array optional. The ids of this cluster's organizations.

desired_instances
integer optional. The number of instances this cluster should attempt to have.

asg_name
string optional. The name of the AutoScaling Group for the Bocce cluster.

queue_url
string optional. The SQS queue URL where jobs are posted to.

Value

A list containing the following elements:

id

integer, The ID of this cluster.

instanceType

string, The EC2 instance types in this cluster.

minInstances

integer, The minimum number of instances in this cluster.

maxInstances

integer, The maximum number of instances in this cluster.

instances

integer, The number of instances currently in this cluster.

instanceMaxMemory

integer, The amount of memory available to a single instance.

instanceMaxCpu

integer, The number of processor shares available to a single instance.

instanceMaxDiskSpace

number, The amount of memory available to a single instance.

region

string, The AWS region that this cluster is in.

activeJobsCount

integer, The number of jobs currently being run in the cluster.

queuedJobsCount

integer, The number of jobs currently waiting to be run on the cluster.

organizationIds

array, The ids of this cluster's organizations.

organizationSlugs

array, The slugs of this cluster's organizations.

singleOrg

boolean, Whether this cluster is allowed to be used by multiple orgs.

desiredInstances

integer, The number of instances this cluster should attempt to have.

remoteDesiredInstances

integer, The number of instances this cluster should attempt to have. Returns -1 if AWS query fails

asgName

string, The name of the AutoScaling Group for the Bocce cluster.

queueUrl

string, The SQS queue URL where jobs are posted to.

createdAt

string, The date and time of cluster creation.

updatedAt

string, The date and time of the cluster's last update.