A table in the Civis Platform

Use civis_table to use a table in the Civis Platform with civis_ml.

civis_table(
  table_name,
  database_name = NULL,
  sql_where = NULL,
  sql_limit = NULL
)

Arguments

table_name

The table name, as "schema.table".

database_name

The name of the database holding table_name.

sql_where

A SQL WHERE clause used to scope the rows of the training or prediction table. Note, the WHERE keyword is added at runtime.

sql_limit

A SQL_LIMIT clause for querying the training or prediction set. Note, the LIMIT keyword is added at runtime.

Value

A civis_table.