List active workloads in an Instance Config

List active workloads in an Instance Config

clusters_list_kubernetes_instance_configs_active_workloads(id, state = NULL)

Arguments

id

integer required. The id of the instance config.

state

string optional. If specified, return workloads in these states. It accepts a comma-separated list, possible values are pending, running

Value

An array containing the following fields:

id

integer, The id of this deployment.

baseType

string, The base type of this deployment.

baseId

integer, The id of the base object associated with this deployment.

baseObjectName

string, The name of the base object associated with this deployment. Null if you do not have permission to read the object.

jobType

string, If the base object is a job run you have permission to read, the type of the job. One of "python_script", "r_script", "container_script", or "custom_script".

jobId

integer, If the base object is a job run you have permission to read, the id of the job.

jobCancelRequestedAt

string, If the base object is a job run you have permission to read, and it was requested to be cancelled, the timestamp of that request.

state

string, The state of this deployment.

cpu

integer, The CPU in millicores requested by this deployment.

memory

integer, The memory in MB requested by this deployment.

diskSpace

integer, The disk space in GB requested by this deployment.

user

list, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

createdAt

string, The timestamp of when the deployment began.

cancellable

boolean, True if you have permission to cancel this deployment.