Update some attributes of this Identity Resolution Enhancement

Update some attributes of this Identity Resolution Enhancement

enhancements_patch_identity_resolution(
  id,
  name = NULL,
  schedule = NULL,
  notifications = NULL,
  threshold = NULL,
  sources = NULL,
  match_target_id = NULL,
  enforced_links = NULL,
  customer_graph = NULL,
  golden_table = NULL,
  link_scores = NULL
)

Arguments

id

integer required. The ID for the enhancement.

name

string optional. The name of the enhancement job.

schedule

list optional. A list containing the following elements:

scheduled boolean, If the item is scheduled.

scheduledDays array, Days of the week, based on numeric value starting at 0 for Sunday. Mutually exclusive with scheduledDaysOfMonth

scheduledHours array, Hours of the day it is scheduled on.

scheduledMinutes array, Minutes of the day it is scheduled on.

scheduledRunsPerHour integer, Deprecated in favor of scheduled minutes.

scheduledDaysOfMonth array, Days of the month it is scheduled on, mutually exclusive with scheduledDays.

notifications

list optional. A list containing the following elements:

urls array, URLs to receive a POST request at job completion

successEmailSubject string, Custom subject line for success e-mail.

successEmailBody string, Custom body text for success e-mail, written in Markdown.

successEmailAddresses array, Addresses to notify by e-mail when the job completes successfully.

successEmailFromName string, Name from which success emails are sent; defaults to "Civis."

successEmailReplyTo string, Address for replies to success emails; defaults to the author of the job.

failureEmailAddresses array, Addresses to notify by e-mail when the job fails.

stallWarningMinutes integer, Stall warning emails will be sent after this amount of minutes.

successOn boolean, If success email notifications are on. Defaults to user's preferences.

failureOn boolean, If failure email notifications are on. Defaults to user's preferences.

threshold

number optional. A value that determines the extent to which similar records get assigned the same resolved ID. Must be within 0.5 and 1, inclusive. Defaults to 0.8 if unspecified.Higher values may result in fewer cases where records about different individuals erroneously receive the same resolved ID, but also more more cases where records about the same individual receive different resolved IDs.

sources

array optional. An array containing the following fields:

name string, A user-specified name for the source.

description string, A description of the source.

databaseName string, The name of the source database.

schemaName string, The name of the source schema.

tableName string, The name of the source table.

fieldMapping object, A mapping of PII fields to columns in this table. Valid keys are primary_key, first_name, middle_name, last_name, gender, phone, email, birth_date, birth_year, birth_month, birth_day, house_number, street, unit, full_address, city, state, state_code, zip, lat, lon, and name_suffix

match_target_id

integer optional. The ID of the Civis Data (Custom) match target. See /match_targets for IDs.

enforced_links

array optional. An array containing the following fields:

source1 string, Name of the first source. Must be defined in Sources list.

source1JoinCol string, Column from the first source to join on.

source2 string, Name of the second source. Must be defined in Sources list

source2JoinCol string, Column from the second source to join on.

customer_graph

list optional. A list containing the following elements:

databaseName string, The name of the destination database.

schemaName string, The name of the destination schema.

tableName string, The name of the destination table.

golden_table

list optional. A list containing the following elements:

databaseName string, The name of the destination database.

schemaName string, The name of the destination schema.

tableName string, The name of the destination table.

fields array,

link_scores

list optional. A list containing the following elements:

databaseName string, The name of the destination database.

schemaName string, The name of the destination schema.

tableName string, The name of the destination table.

Value

A list containing the following elements:

id

integer, The ID for the enhancement.

name

string, The name of the enhancement job.

type

string, The type of the enhancement (e.g CASS-NCOA)

createdAt

string, The time this enhancement was created.

updatedAt

string, The time the enhancement was last updated.

author

list, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

state

string, The status of the enhancement's last run

schedule

list, A list containing the following elements:

scheduled boolean, If the item is scheduled.

scheduledDays array, Days of the week, based on numeric value starting at 0 for Sunday. Mutually exclusive with scheduledDaysOfMonth

scheduledHours array, Hours of the day it is scheduled on.

scheduledMinutes array, Minutes of the day it is scheduled on.

scheduledRunsPerHour integer, Deprecated in favor of scheduled minutes.

scheduledDaysOfMonth array, Days of the month it is scheduled on, mutually exclusive with scheduledDays.

notifications

list, A list containing the following elements:

urls array, URLs to receive a POST request at job completion

successEmailSubject string, Custom subject line for success e-mail.

successEmailBody string, Custom body text for success e-mail, written in Markdown.

successEmailAddresses array, Addresses to notify by e-mail when the job completes successfully.

successEmailFromName string, Name from which success emails are sent; defaults to "Civis."

successEmailReplyTo string, Address for replies to success emails; defaults to the author of the job.

failureEmailAddresses array, Addresses to notify by e-mail when the job fails.

stallWarningMinutes integer, Stall warning emails will be sent after this amount of minutes.

successOn boolean, If success email notifications are on. Defaults to user's preferences.

failureOn boolean, If failure email notifications are on. Defaults to user's preferences.

runningAs

list, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

myPermissionLevel

string, Your permission level on the object. One of "read", "write", or "manage".

archived

string, The archival status of the requested item(s).

threshold

number, A value that determines the extent to which similar records get assigned the same resolved ID. Must be within 0.5 and 1, inclusive. Defaults to 0.8 if unspecified.Higher values may result in fewer cases where records about different individuals erroneously receive the same resolved ID, but also more more cases where records about the same individual receive different resolved IDs.

sources

array, An array containing the following fields:

name string, A user-specified name for the source.

description string, A description of the source.

databaseName string, The name of the source database.

schemaName string, The name of the source schema.

tableName string, The name of the source table.

fieldMapping object, A mapping of PII fields to columns in this table. Valid keys are primary_key, first_name, middle_name, last_name, gender, phone, email, birth_date, birth_year, birth_month, birth_day, house_number, street, unit, full_address, city, state, state_code, zip, lat, lon, and name_suffix

matchTargetId

integer, The ID of the Civis Data (Custom) match target. See /match_targets for IDs.

enforcedLinks

array, An array containing the following fields:

source1 string, Name of the first source. Must be defined in Sources list.

source1JoinCol string, Column from the first source to join on.

source2 string, Name of the second source. Must be defined in Sources list

source2JoinCol string, Column from the second source to join on.

customerGraph

list, A list containing the following elements:

databaseName string, The name of the destination database.

schemaName string, The name of the destination schema.

tableName string, The name of the destination table.

goldenTable

list, A list containing the following elements:

databaseName string, The name of the destination database.

schemaName string, The name of the destination schema.

tableName string, The name of the destination table.

fields array,

linkScores

list, A list containing the following elements:

databaseName string, The name of the destination database.

schemaName string, The name of the destination schema.

tableName string, The name of the destination table.

legacyId

integer, ID of this pipeline in the legacy IDR service application.

lastRun

list, A list containing the following elements:

id integer,

state string,

createdAt string, The time that the run was queued.

startedAt string, The time that the run started.

finishedAt string, The time that the run completed.

error string, The error message for this run, if present.

config string, How the Identity Resolution job was configured for this run.

sampleRecordsQuery string, A SQL query to produce a sample of records to inspect.

expandClusterQuery string, A customizable query to view PII associated with resolved ids.

runMetrics list . A list containing the following elements:

numRecords integer, The number of input records for this run.

uniqueIds integer, The number of distinct unique IDs in the input records for this run.

uniqueDeduplicatedIds integer, The number of resolved IDs associated with more than one unique ID in the input.

maxClusterSize integer, The number of records in the largest cluster of resolved IDs.

avgClusterSize number, The average number of records with the same resolved ID.

clusterSizeFrequencies object, A mapping from numbers of records with the same resolved ID (i.e., sizes of clusters) to numbers of such clusters. For example, if there were 10 clusters with 2 records each, 2 would be a key in the mapping, and 10 would be its value.

errorSection string, If there was a failure, this will denote which section of the Identity Resolution job failed. One of: data_preparation, compute_setup or data_processing.