Create a CASS/NCOA Enhancement

Create a CASS/NCOA Enhancement

enhancements_post_cass_ncoa(name, source, schedule = NULL, parent_id = NULL,
  notifications = NULL, destination = NULL, column_mapping = NULL,
  use_default_column_mapping = NULL, perform_ncoa = NULL,
  ncoa_credential_id = NULL, output_level = NULL, limiting_sql = NULL)

Arguments

name
string required. The name of the enhancement.

source
object required. A list containing the following elements:

databaseTable object,

schedule
object optional. A list containing the following elements:

scheduled boolean, If the object is scheduled

scheduledDays array, Day based on numeric value starting at 0 for Sunday

scheduledHours array, Hours of the day it is scheduled on

scheduledMinutes array, Minutes of the day it is scheduled on

scheduledRunsPerHour integer, Alternative to scheduled minutes, number of times to run per hour

parent_id
integer optional. Parent ID that triggers this enhancement.

notifications
object optional. A list containing the following elements:

urls array, URLs to receive a POST request at job completion

successEmailSubject string, Custom subject line for success e-mail.

successEmailBody string, Custom body text for success e-mail, written in Markdown.

successEmailAddresses array, Addresses to notify by e-mail when the job completes successfully.

failureEmailAddresses array, Addresses to notify by e-mail when the job fails.

stallWarningMinutes integer, Stall warning emails will be sent after this amount of minutes.

successOn boolean, If success email notifications are on

failureOn boolean, If failure email notifications are on

destination
object optional. A list containing the following elements:

databaseTable object,

column_mapping
object optional. A list containing the following elements:

address1 string, The first address line.

address2 string, The second address line.

city string, The city of an address.

state string, The state of an address.

zip string, The zip code of an address.

name string, The full name of the resident at this address. If needed, separate multiple columns with `+`, e.g. `first_name+last_name`

company string, The name of the company located at this address.

use_default_column_mapping
boolean optional. Defaults to true, where the existing column mapping on the input table will be used. If false, a custom column mapping must be provided.

perform_ncoa
boolean optional. Whether to update addresses for records matching the National Change of Address (NCOA) database.

ncoa_credential_id
integer optional. Credential to use when performing NCOA updates. Required if 'performNcoa' is true.

output_level
string optional. The set of fields persisted by a CASS or NCOA enhancement.For CASS enhancements, one of 'cass' or 'all.'For NCOA enhancements, one of 'cass', 'ncoa' , 'coalesced' or 'all'.By default, all fields will be returned.

limiting_sql
string optional. The limiting sql.

Value

A list containing the following elements:

id

integer, The ID for the enhancement.

name

string, The name of the enhancement.

type

string, The type of the enhancement (e.g CASS-NCOA)

createdAt

string, The time this enhancement was created.

updatedAt

string, The time the enhancement was last updated.

author

object, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

state

string, The status of the enhancement's last run

schedule

object, A list containing the following elements:

scheduled boolean, If the object is scheduled

scheduledDays array, Day based on numeric value starting at 0 for Sunday

scheduledHours array, Hours of the day it is scheduled on

scheduledMinutes array, Minutes of the day it is scheduled on

scheduledRunsPerHour integer, Alternative to scheduled minutes, number of times to run per hour

parentId

integer, Parent ID that triggers this enhancement.

notifications

object, A list containing the following elements:

urls array, URLs to receive a POST request at job completion

successEmailSubject string, Custom subject line for success e-mail.

successEmailBody string, Custom body text for success e-mail, written in Markdown.

successEmailAddresses array, Addresses to notify by e-mail when the job completes successfully.

failureEmailAddresses array, Addresses to notify by e-mail when the job fails.

stallWarningMinutes integer, Stall warning emails will be sent after this amount of minutes.

successOn boolean, If success email notifications are on

failureOn boolean, If failure email notifications are on

runningAs

object, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

source

object, A list containing the following elements:

databaseTable object,

destination

object, A list containing the following elements:

databaseTable object,

columnMapping

object, A list containing the following elements:

address1 string, The first address line.

address2 string, The second address line.

city string, The city of an address.

state string, The state of an address.

zip string, The zip code of an address.

name string, The full name of the resident at this address. If needed, separate multiple columns with `+`, e.g. `first_name+last_name`

company string, The name of the company located at this address.

useDefaultColumnMapping

boolean, Defaults to true, where the existing column mapping on the input table will be used. If false, a custom column mapping must be provided.

performNcoa

boolean, Whether to update addresses for records matching the National Change of Address (NCOA) database.

ncoaCredentialId

integer, Credential to use when performing NCOA updates. Required if 'performNcoa' is true.

outputLevel

string, The set of fields persisted by a CASS or NCOA enhancement.For CASS enhancements, one of 'cass' or 'all.'For NCOA enhancements, one of 'cass', 'ncoa' , 'coalesced' or 'all'.By default, all fields will be returned.

limitingSQL

string, The limiting sql.

archived

string, The archival status of the requested object(s).