Replace all attributes of this Geocode Enhancement

Replace all attributes of this Geocode Enhancement

enhancements_put_geocode(
  id,
  name,
  remote_host_id,
  credential_id,
  source_schema_and_table,
  schedule = NULL,
  parent_id = NULL,
  notifications = NULL,
  multipart_key = NULL,
  limiting_sql = NULL,
  target_schema = NULL,
  target_table = NULL,
  country = NULL,
  provider = NULL,
  output_address = NULL
)

Arguments

id
integer required. The ID for the enhancement.

name
string required. The name of the enhancement job.

remote_host_id
integer required. The ID of the remote host.

credential_id
integer required. The ID of the remote host credential.

source_schema_and_table
string required. The source database schema and table.

schedule
list optional. A list containing the following elements:

scheduled boolean, If the item is scheduled.

scheduledDays array, Day based on numeric value starting at 0 for Sunday.

scheduledHours array, Hours of the day it is scheduled on.

scheduledMinutes array, Minutes of the day it is scheduled on.

scheduledRunsPerHour integer, Alternative to scheduled minutes, number of times to run per hour.

parent_id
integer optional. Parent ID that triggers this enhancement.

notifications
list optional. A list containing the following elements:

urls array, URLs to receive a POST request at job completion

successEmailSubject string, Custom subject line for success e-mail.

successEmailBody string, Custom body text for success e-mail, written in Markdown.

successEmailAddresses array, Addresses to notify by e-mail when the job completes successfully.

successEmailFromName string, Name from which success emails are sent; defaults to "Civis."

successEmailReplyTo string, Address for replies to success emails; defaults to the author of the job.

failureEmailAddresses array, Addresses to notify by e-mail when the job fails.

stallWarningMinutes integer, Stall warning emails will be sent after this amount of minutes.

successOn boolean, If success email notifications are on.

failureOn boolean, If failure email notifications are on.

multipart_key
array optional. The source table primary key.

limiting_sql
string optional. The limiting SQL for the source table. "WHERE" should be omitted (e.g. state='IL').

target_schema
string optional. The output table schema.

target_table
string optional. The output table name.

country
string optional. The country of the addresses to be geocoded; either 'us' or 'ca'.

provider
string optional. The geocoding provider; one of postgis, nominatim, and geocoder_ca.

output_address
boolean optional. Whether to output the parsed address. Only guaranteed for the 'postgis' provider.

Value

A list containing the following elements:

id

integer, The ID for the enhancement.

name

string, The name of the enhancement job.

type

string, The type of the enhancement (e.g CASS-NCOA)

createdAt

string, The time this enhancement was created.

updatedAt

string, The time the enhancement was last updated.

author

list, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

state

string, The status of the enhancement's last run

schedule

list, A list containing the following elements:

scheduled boolean, If the item is scheduled.

scheduledDays array, Day based on numeric value starting at 0 for Sunday.

scheduledHours array, Hours of the day it is scheduled on.

scheduledMinutes array, Minutes of the day it is scheduled on.

scheduledRunsPerHour integer, Alternative to scheduled minutes, number of times to run per hour.

parentId

integer, Parent ID that triggers this enhancement.

notifications

list, A list containing the following elements:

urls array, URLs to receive a POST request at job completion

successEmailSubject string, Custom subject line for success e-mail.

successEmailBody string, Custom body text for success e-mail, written in Markdown.

successEmailAddresses array, Addresses to notify by e-mail when the job completes successfully.

successEmailFromName string, Name from which success emails are sent; defaults to "Civis."

successEmailReplyTo string, Address for replies to success emails; defaults to the author of the job.

failureEmailAddresses array, Addresses to notify by e-mail when the job fails.

stallWarningMinutes integer, Stall warning emails will be sent after this amount of minutes.

successOn boolean, If success email notifications are on.

failureOn boolean, If failure email notifications are on.

runningAs

list, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

remoteHostId

integer, The ID of the remote host.

credentialId

integer, The ID of the remote host credential.

sourceSchemaAndTable

string, The source database schema and table.

multipartKey

array, The source table primary key.

limitingSQL

string, The limiting SQL for the source table. "WHERE" should be omitted (e.g. state='IL').

targetSchema

string, The output table schema.

targetTable

string, The output table name.

country

string, The country of the addresses to be geocoded; either 'us' or 'ca'.

provider

string, The geocoding provider; one of postgis, nominatim, and geocoder_ca.

outputAddress

boolean, Whether to output the parsed address. Only guaranteed for the 'postgis' provider.

archived

string, The archival status of the requested item(s).