Update the archive status of this object

Update the archive status of this object

enhancements_put_person_matching_archive(id, status)

Arguments

id
integer required. The ID of the object.

status
boolean required. The desired archived status of the object.

Value

A list containing the following elements:

id

integer, The ID for the enhancement.

name

string, The name of the enhancement job.

type

string, The type of the enhancement (e.g CASS-NCOA)

createdAt

string, The time this enhancement was created.

updatedAt

string, The time the enhancement was last updated.

author

list, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

state

string, The status of the enhancement's last run

schedule

list, A list containing the following elements:

scheduled boolean, If the item is scheduled.

scheduledDays array, Day based on numeric value starting at 0 for Sunday.

scheduledHours array, Hours of the day it is scheduled on.

scheduledMinutes array, Minutes of the day it is scheduled on.

scheduledRunsPerHour integer, Alternative to scheduled minutes, number of times to run per hour.

parentId

integer, Parent ID that triggers this enhancement.

notifications

list, A list containing the following elements:

urls array, URLs to receive a POST request at job completion

successEmailSubject string, Custom subject line for success e-mail.

successEmailBody string, Custom body text for success e-mail, written in Markdown.

successEmailAddresses array, Addresses to notify by e-mail when the job completes successfully.

successEmailFromName string, Name from which success emails are sent; defaults to "Civis."

successEmailReplyTo string, Address for replies to success emails; defaults to the author of the job.

failureEmailAddresses array, Addresses to notify by e-mail when the job fails.

stallWarningMinutes integer, Stall warning emails will be sent after this amount of minutes.

successOn boolean, If success email notifications are on.

failureOn boolean, If failure email notifications are on.

runningAs

list, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

configuration

list, A list containing the following elements:

task string, The type of person matching task. Options are: "table_to_table", "dedupe_table", or "table_to_civis_data".

source string, The input source of your data. Options are: "redshift" or "s3".

inputDatabaseName string, The Redshift database name for input data.

inputSchema string, The schema name for the input data.

inputTable string, The table name for the input data.

inputFileId string, The ID of the input S3 file.

inputFieldMapping string, The column mapping for the input in JSON or YAML.

inputFileHeaders string, Provide your headers in a list if the first row of your input does not have the headers, and make them JSON-decodable. For example: ["col1","col2","col3"].

targetDatabaseName string, The Redshift database for target data.

targetSchema string, The schema for target data.

targetTable string, The table for target data.

targetFieldMapping string, The column mapping for the target in JSON or YAML.

targetFileId string, The ID of the target S3 file.

matchTargetId integer, The ID of the match target.

matchDatabaseName string, The Redshift database for the match output table.

matchSchema string, The schema for the match output table.

matchTable string, The name of the match output table.

matchCsvFilename string, The name of the match output file.

matchSourceIdCol string, The name of the column in the output table that will hold the id from the source for each match.

matchTargetIdCol string, The name the column in the output table that will hold the id from the target for each match.

maxMatches integer, The maximum number of matches to return.

threshold number, The score threshold (between 0 and 1).