Get a CSV Export

Get a CSV Export

exports_get_files_csv(id)

Arguments

id
integer required.

Value

A list containing the following elements:

id

integer, The ID of this Csv Export job.

name

string, The name of this Csv Export job.

source

list, A list containing the following elements:

sql string, The SQL query for this Csv Export job

remoteHostId integer, The ID of the destination database host.

credentialId integer, The ID of the credentials for the destination database.

destination

list, A list containing the following elements:

filenamePrefix string, The prefix of the name of the file returned to the user.

storagePath list . A list containing the following elements:

filePath string, The path within the bucket where the exported file will be saved. E.g. the file_path for "s3://mybucket/files/all/" would be "/files/all/"

storageHostId integer, The ID of the destination storage host.

credentialId integer, The ID of the credentials for the destination storage host.

existingFiles string, Notifies the job of what to do in the case that the exported file already exists at the provided path.One of: fail, append, overwrite. Default: fail. If "append" is specified,the new file will always be added to the provided path. If "overwrite" is specifiedall existing files at the provided path will be deleted and the new file will be added.By default, or if "fail" is specified, the export will fail if a file exists at the provided path.

includeHeader

boolean, A boolean value indicating whether or not the header should be included. Defaults to true.

compression

string, The compression of the output file. Valid arguments are "gzip" and "none". Defaults to "gzip".

columnDelimiter

string, The column delimiter for the output file. Valid arguments are "comma", "tab", and "pipe". Defaults to "comma".

hidden

boolean, A boolean value indicating whether or not this request should be hidden. Defaults to false.

forceMultifile

boolean, Whether or not the csv should be split into multiple files. Default: false

maxFileSize

integer, The max file size, in MB, created files will be. Only available when force_multifile is true.