Create a Preprocess CSV

Create a Preprocess CSV

files_post_preprocess_csv(
  file_id,
  in_place = NULL,
  detect_table_columns = NULL,
  force_character_set_conversion = NULL,
  include_header = NULL,
  column_delimiter = NULL,
  hidden = NULL
)

Arguments

file_id
integer required. The ID of the file.

in_place
boolean optional. If true, the file is cleaned in place. If false, a new file ID is created. Defaults to true.

detect_table_columns
boolean optional. If true, detect the table columns in the file including the sql types. If false, skip table column detection.Defaults to false.

force_character_set_conversion
boolean optional. If true, the file will always be converted to UTF-8 and any character that cannot be converted will be discarded. If false, the character set conversion will only run if the detected character set is not compatible with UTF-8 (e.g., UTF-8, ASCII).

include_header
boolean optional. A boolean value indicating whether or not the first row of the file is a header row. If not provided, will attempt to auto-detect whether a header row is present.

column_delimiter
string optional. The column delimiter for the file. One of "comma", "tab", or "pipe". If not provided, the column delimiter will be auto-detected.

hidden
boolean optional. The hidden status of the item.

Value

A list containing the following elements:

id

integer, The ID of the job created.

fileId

integer, The ID of the file.

inPlace

boolean, If true, the file is cleaned in place. If false, a new file ID is created. Defaults to true.

detectTableColumns

boolean, If true, detect the table columns in the file including the sql types. If false, skip table column detection.Defaults to false.

forceCharacterSetConversion

boolean, If true, the file will always be converted to UTF-8 and any character that cannot be converted will be discarded. If false, the character set conversion will only run if the detected character set is not compatible with UTF-8 (e.g., UTF-8, ASCII).

includeHeader

boolean, A boolean value indicating whether or not the first row of the file is a header row. If not provided, will attempt to auto-detect whether a header row is present.

columnDelimiter

string, The column delimiter for the file. One of "comma", "tab", or "pipe". If not provided, the column delimiter will be auto-detected.

hidden

boolean, The hidden status of the item.