Get a script function matching a job type.

Get a script function matching a job type.

get_script_fun(job, verb = c("list", "post"), fun_type = c("outputs", "logs"))

Arguments

job

output of jobs_get

verb

one of "list" or "post"

fun_type

one of "logs" or "outputs"

Value

The correct output or log fetching function (e.g. scripts_list_containers_runs_logs) based on the job type.

Details

container and custom scripts both have the same job type, but can be distinguished by a non-null fromTemplateId.