Replace all attributes of this Group

Replace all attributes of this Group

groups_put(
  id,
  name,
  description = NULL,
  slug = NULL,
  organization_id = NULL,
  must_agree_to_eula = NULL,
  default_otp_required_for_login = NULL,
  role_ids = NULL,
  default_time_zone = NULL,
  default_jobs_label = NULL,
  default_notebooks_label = NULL,
  default_services_label = NULL
)

Arguments

id
integer required. The ID of this group.

name
string required. This group's name.

description
string optional. The description of the group.

slug
string optional. The slug for this group.

organization_id
integer optional. The ID of the organization this group belongs to.

must_agree_to_eula
boolean optional. Whether or not members of this group must sign the EULA.

default_otp_required_for_login
boolean optional. The two factor authentication requirement for this group.

role_ids
array optional. An array of ids of all the roles this group has.

default_time_zone
string optional. The default time zone of this group.

default_jobs_label
string optional. The default partition label for jobs of this group. Only available if custom_partitions feature flag is set. Do not use this attribute as it may break in the future.

default_notebooks_label
string optional. The default partition label for notebooks of this group. Only available if custom_partitions feature flag is set. Do not use this attribute as it may break in the future.

default_services_label
string optional. The default partition label for services of this group. Only available if custom_partitions feature flag is set. Do not use this attribute as it may break in the future.

Value

A list containing the following elements:

id

integer, The ID of this group.

name

string, This group's name.

createdAt

string, The date and time when this group was created.

description

string, The description of the group.

slug

string, The slug for this group.

organizationId

integer, The ID of the organization this group belongs to.

organizationName

string, The name of the organization this group belongs to.

memberCount

integer, The total number of members in this group.

mustAgreeToEula

boolean, Whether or not members of this group must sign the EULA.

defaultOtpRequiredForLogin

boolean, The two factor authentication requirement for this group.

roleIds

array, An array of ids of all the roles this group has.

defaultTimeZone

string, The default time zone of this group.

defaultJobsLabel

string, The default partition label for jobs of this group. Only available if custom_partitions feature flag is set. Do not use this attribute as it may break in the future.

defaultNotebooksLabel

string, The default partition label for notebooks of this group. Only available if custom_partitions feature flag is set. Do not use this attribute as it may break in the future.

defaultServicesLabel

string, The default partition label for services of this group. Only available if custom_partitions feature flag is set. Do not use this attribute as it may break in the future.

members

array, An array containing the following fields:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.