List Imports

List Imports

imports_list(
  type = NULL,
  destination = NULL,
  source = NULL,
  status = NULL,
  author = NULL,
  hidden = NULL,
  archived = NULL,
  limit = NULL,
  page_num = NULL,
  order = NULL,
  order_dir = NULL
)

Arguments

type

string optional. If specified, return imports of these types. It accepts a comma-separated list, possible values are Dbsync, AutoImport, GdocImport, and GdocExport.

destination

string optional. If specified, returns imports with one of these destinations. It accepts a comma-separated list of remote host ids.

source

string optional. If specified, returns imports with one of these sources. It accepts a comma-separated list of remote host ids. 'Dbsync' must be specified for 'type'.

status

string optional. If specified, returns imports with one of these statuses. It accepts a comma-separated list, possible values are 'running', 'failed', 'succeeded', 'idle', 'scheduled'.

author

string optional. If specified, return items from any of these authors. It accepts a comma-separated list of user IDs.

hidden

boolean optional. If specified to be true, returns hidden items. Defaults to false, returning non-hidden items.

archived

string optional. The archival status of the requested item(s).

limit

integer optional. Number of results to return. Defaults to 20. Maximum allowed is 50.

page_num

integer optional. Page number of the results to return. Defaults to the first page, 1.

order

string optional. The field on which to order the result set. Defaults to updated_at. Must be one of: updated_at, name, created_at, last_run.updated_at.

order_dir

string optional. Direction in which to sort, either asc (ascending) or desc (descending) defaulting to desc.

Value

An array containing the following fields:

name

string, The name of the import.

syncType

string, The type of sync to perform; one of Dbsync, AutoImport, GdocImport, and GdocExport.

source

list, A list containing the following elements:

remoteHostId integer,

credentialId integer,

additionalCredentials array, Array that holds additional credentials used for specific imports. For DB Syncs, the first element is an SSL private key credential id, and the second element is the corresponding public key credential id.

name string,

destination

list, A list containing the following elements:

remoteHostId integer,

credentialId integer,

additionalCredentials array, Array that holds additional credentials used for specific imports. For DB Syncs, the first element is an SSL private key credential id, and the second element is the corresponding public key credential id.

name string,

schedule

list, A list containing the following elements:

scheduled boolean, If the item is scheduled.

scheduledDays array, Days of the week, based on numeric value starting at 0 for Sunday. Mutually exclusive with scheduledDaysOfMonth

scheduledHours array, Hours of the day it is scheduled on.

scheduledMinutes array, Minutes of the day it is scheduled on.

scheduledRunsPerHour integer, Deprecated in favor of scheduled minutes.

scheduledDaysOfMonth array, Days of the month it is scheduled on, mutually exclusive with scheduledDays.

id

integer, The ID for the import.

isOutbound

boolean,

jobType

string, The job type of this import.

state

string,

createdAt

string,

updatedAt

string,

lastRun

list, A list containing the following elements:

id integer,

state string,

createdAt string, The time that the run was queued.

startedAt string, The time that the run started.

finishedAt string, The time that the run completed.

error string, The error message for this run, if present.

user

list, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

timeZone

string, The time zone of this import.

archived

string, The archival status of the requested item(s).