Create a new import configuration

Create a new import configuration

imports_post(name, sync_type, is_outbound, source = NULL,
  destination = NULL, schedule = NULL, notifications = NULL,
  parent_id = NULL, next_run_at = NULL, time_zone = NULL, hidden = NULL)

Arguments

name
string required. The name of the import.

sync_type
string required. The type of sync to perform; one of Dbsync, AutoImport, SilverpopDataImport, SilverpopContactImport, GdocImport, GdocExport, and Salesforce.

is_outbound
boolean required.

source
object optional. A list containing the following elements:

remoteHostId integer,

credentialId integer,

additionalCredentials array, Array that holds additional credentials used for specific imports. For salesforce imports, the first and only element is the client credential id.

destination
object optional. A list containing the following elements:

remoteHostId integer,

credentialId integer,

additionalCredentials array, Array that holds additional credentials used for specific imports. For salesforce imports, the first and only element is the client credential id.

schedule
object optional. A list containing the following elements:

scheduled boolean, If the object is scheduled

scheduledDays array, Day based on numeric value starting at 0 for Sunday

scheduledHours array, Hours of the day it is scheduled on

scheduledMinutes array, Minutes of the day it is scheduled on

scheduledRunsPerHour integer, Alternative to scheduled minutes, number of times to run per hour

notifications
object optional. A list containing the following elements:

urls array, URLs to receive a POST request at job completion

successEmailSubject string, Custom subject line for success e-mail.

successEmailBody string, Custom body text for success e-mail, written in Markdown.

successEmailAddresses array, Addresses to notify by e-mail when the job completes successfully.

failureEmailAddresses array, Addresses to notify by e-mail when the job fails.

stallWarningMinutes integer, Stall warning emails will be sent after this amount of minutes.

successOn boolean, If success email notifications are on

failureOn boolean, If failure email notifications are on

parent_id
integer optional. Parent id to trigger this import from

next_run_at
string optional. The time of the next scheduled run.

time_zone
string optional. The time zone of this import.

hidden
boolean optional. The hidden status of the object. Setting this to true hides it from most API endpoints. The object can still be queried directly by ID

Value

A list containing the following elements:

name

string, The name of the import.

syncType

string, The type of sync to perform; one of Dbsync, AutoImport, SilverpopDataImport, SilverpopContactImport, GdocImport, GdocExport, and Salesforce.

source

object, A list containing the following elements:

remoteHostId integer,

credentialId integer,

additionalCredentials array, Array that holds additional credentials used for specific imports. For salesforce imports, the first and only element is the client credential id.

name string,

destination

object, A list containing the following elements:

remoteHostId integer,

credentialId integer,

additionalCredentials array, Array that holds additional credentials used for specific imports. For salesforce imports, the first and only element is the client credential id.

name string,

schedule

object, A list containing the following elements:

scheduled boolean, If the object is scheduled

scheduledDays array, Day based on numeric value starting at 0 for Sunday

scheduledHours array, Hours of the day it is scheduled on

scheduledMinutes array, Minutes of the day it is scheduled on

scheduledRunsPerHour integer, Alternative to scheduled minutes, number of times to run per hour

notifications

object, A list containing the following elements:

urls array, URLs to receive a POST request at job completion

successEmailSubject string, Custom subject line for success e-mail.

successEmailBody string, Custom body text for success e-mail, written in Markdown.

successEmailAddresses array, Addresses to notify by e-mail when the job completes successfully.

failureEmailAddresses array, Addresses to notify by e-mail when the job fails.

stallWarningMinutes integer, Stall warning emails will be sent after this amount of minutes.

successOn boolean, If success email notifications are on

failureOn boolean, If failure email notifications are on

parentId

integer, Parent id to trigger this import from

id

integer, The ID for the import.

isOutbound

boolean,

syncs

array, An array containing the following fields:

id integer,

source object,

destination object,

advancedOptions object,

state

string,

createdAt

string,

updatedAt

string,

lastRun

object, A list containing the following elements:

id integer,

state string,

createdAt string, The time that the run was queued.

startedAt string, The time that the run started.

finishedAt string, The time that the run completed.

error string, The error message for this run, if present.

user

object, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

runningAs

object, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

nextRunAt

string, The time of the next scheduled run.

timeZone

string, The time zone of this import.

hidden

boolean, The hidden status of the object. Setting this to true hides it from most API endpoints. The object can still be queried directly by ID

archived

string, The archival status of the requested object(s).