List

List

jobs_list(state = NULL, type = NULL, q = NULL, permission = NULL,
  scheduled = NULL, hidden = NULL, archived = NULL, limit = NULL,
  page_num = NULL, order = NULL, order_dir = NULL)

Arguments

state
string optional. The job's state. One or more of queued, running, succeeded, failed, and cancelled. Specify multiple values as a comma-separated list (e.g., "A,B").

type
string optional. The job's type. Specify multiple values as a comma-separated list (e.g., "A,B").

q
string optional. Query string to search on the id, name, and job type.

permission
string optional. A permissions string, one of "read", "write", or "manage". Lists only jobs for which the current user has that permission.

scheduled
boolean optional. If the item is scheduled.

hidden
boolean optional. If specified to be true, returns hidden items. Defaults to false, returning non-hidden items.

archived
string optional. The archival status of the requested item(s).

limit
integer optional. Number of results to return. Defaults to its maximum of 50.

page_num
integer optional. Page number of the results to return. Defaults to the first page, 1.

order
string optional. The field on which to order the result set. Defaults to updated_at. Must be one of: updated_at.

order_dir
string optional. Direction in which to sort, either asc (ascending) or desc (descending) defaulting to desc.

Value

An array containing the following fields:

id

integer,

name

string,

type

string,

fromTemplateId

integer,

state

string, Whether the job is idle, queued, running, cancelled, or failed.

createdAt

string,

updatedAt

string,

lastRun

list, A list containing the following elements:

id integer,

state string,

createdAt string, The time that the run was queued.

startedAt string, The time that the run started.

finishedAt string, The time that the run completed.

error string, The error message for this run, if present.

archived

string, The archival status of the requested item(s).

schedule

list, A list containing the following elements:

scheduled boolean, If the item is scheduled.

scheduledDays array, Day based on numeric value starting at 0 for Sunday.

scheduledHours array, Hours of the day it is scheduled on.

scheduledMinutes array, Minutes of the day it is scheduled on.

scheduledRunsPerHour integer, Alternative to scheduled minutes, number of times to run per hour.