Edit an existing optimization

Edit an existing optimization

media_patch_optimizations(
  id,
  name = NULL,
  runs = NULL,
  programs = NULL,
  networks = NULL,
  exclude_programs = NULL,
  exclude_networks = NULL,
  time_slot_percentages = NULL
)

Arguments

id

integer required. The optimization ID.

name

string optional. The name of the optimization.

runs

array optional. An array containing the following fields:

marketId integer, The market ID.

startDate string, The start date for the media run.

endDate string, The end date for the media run.

forceCpm boolean, Whether to force optimization to use CPM data even if partition data is available.

reachAlpha number, A tuning parameter used to adjust RF.

syscodes array, The syscodes for the media run.

rateCards array, The ratecards for the media run.

constraints array, The constraints for the media run.

programs

array optional. An array of programs that the Civis Media Optimizer either exclude or limit to.An error will be thrown if exclude_programs is not also set.

networks

array optional. An array of networks that the Civis Media Optimizer either exclude or limit to.An error will be thrown if exclude_networks is not also set.

exclude_programs

boolean optional. If Civis Media Optimizer should exclude the programs in the programs parameter.If this value is set to false, it will make the optimization limit itself to the programs supplied through the programs parameter.An error will be thrown if programs is not also set.

exclude_networks

boolean optional. If Civis Media Optimizer should exclude the networks in the networks parameter.If this value is set to false, it will make the optimization limit itself to the networks supplied through the networks.An error will be thrown if networks is not also set.

time_slot_percentages

list optional. The maximum amount of the budget spent on that particular day of the week, daypart, or specific time slot for broadcast and cable.

Value

A list containing the following elements:

id

integer, The optimization ID.

author

list, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

name

string, The name of the optimization.

createdAt

string,

updatedAt

string,

finishedAt

string, The end time of the last run.

state

string, The state of the last run.

lastRunId

integer, The ID of the last run.

spotOrderId

integer, The ID for the spot order produced by the optimization.

archived

string, The archival status of the requested item(s).

reportLink

string, A link to the visual report for the optimization.

spotOrderLink

string, A link to the json version of the spot order.

fileLinks

array, Links to the csv and xml versions of the spot order.

runs

array, An array containing the following fields:

marketId integer, The market ID.

startDate string, The start date for the media run.

endDate string, The end date for the media run.

forceCpm boolean, Whether to force optimization to use CPM data even if partition data is available.

reachAlpha number, A tuning parameter used to adjust RF.

syscodes array, The syscodes for the media run.

rateCards array, The ratecards for the media run.

constraints array, The constraints for the media run.

programs

array, An array of programs that the Civis Media Optimizer either exclude or limit to.An error will be thrown if exclude_programs is not also set.

networks

array, An array of networks that the Civis Media Optimizer either exclude or limit to.An error will be thrown if exclude_networks is not also set.

excludePrograms

boolean, If Civis Media Optimizer should exclude the programs in the programs parameter.If this value is set to false, it will make the optimization limit itself to the programs supplied through the programs parameter.An error will be thrown if programs is not also set.

excludeNetworks

boolean, If Civis Media Optimizer should exclude the networks in the networks parameter.If this value is set to false, it will make the optimization limit itself to the networks supplied through the networks.An error will be thrown if networks is not also set.

timeSlotPercentages

list, The maximum amount of the budget spent on that particular day of the week, daypart, or specific time slot for broadcast and cable.