List

List

models_list(
  model_name = NULL,
  training_table_name = NULL,
  dependent_variable = NULL,
  status = NULL,
  author = NULL,
  hidden = NULL,
  archived = NULL,
  limit = NULL,
  page_num = NULL,
  order = NULL,
  order_dir = NULL
)

Arguments

model_name

string optional. If specified, will be used to filter the models returned. Substring matching is supported. (e.g., "modelName=model" will return both "model1" and "my model").

training_table_name

string optional. If specified, will be used to filter the models returned by the training dataset table name. Substring matching is supported. (e.g., "trainingTableName=table" will return both "table1" and "my_table").

dependent_variable

string optional. If specified, will be used to filter the models returned by the dependent variable column name. Substring matching is supported. (e.g., "dependentVariable=predictor" will return both "predictor" and "my predictor").

status

string optional. If specified, returns models with one of these statuses. It accepts a comma-separated list, possible values are 'running', 'failed', 'succeeded', 'idle', 'scheduled'.

author

string optional. If specified, return items from any of these authors. It accepts a comma-separated list of user IDs.

hidden

boolean optional. If specified to be true, returns hidden items. Defaults to false, returning non-hidden items.

archived

string optional. The archival status of the requested item(s).

limit

integer optional. Number of results to return. Defaults to its maximum of 50.

page_num

integer optional. Page number of the results to return. Defaults to the first page, 1.

order

string optional. The field on which to order the result set. Defaults to updated_at. Must be one of: updated_at, name, created_at, last_run.updated_at.

order_dir

string optional. Direction in which to sort, either asc (ascending) or desc (descending) defaulting to desc.

Value

An array containing the following fields:

id

integer, The ID of the model.

tableName

string, The qualified name of the table containing the training set from which to build the model.

databaseId

integer, The ID of the database holding the training set table used to build the model.

credentialId

integer, The ID of the credential used to read the target table. Defaults to the user's default credential.

modelName

string, The name of the model.

description

string, A description of the model.

interactionTerms

boolean, Whether to search for interaction terms.

boxCoxTransformation

boolean, Whether to transform data so that it assumes a normal distribution. Valid only with continuous models.

modelTypeId

integer, The ID of the model's type.

primaryKey

string, The unique ID (primary key) of the training dataset.

dependentVariable

string, The dependent variable of the training dataset.

dependentVariableOrder

array, The order of dependent variables, especially useful for Ordinal Modeling.

excludedColumns

array, A list of columns which will be considered ineligible to be independent variables.

limitingSQL

string, A custom SQL WHERE clause used to filter the rows used to build the model. (e.g., "id > 105").

crossValidationParameters

list, Cross validation parameter grid for tree methods, e.g. {"n_estimators": [100, 200, 500], "learning_rate": [0.01, 0.1], "max_depth": [2, 3]}.

numberOfFolds

integer, Number of folds for cross validation. Default value is 5.

schedule

list, A list containing the following elements:

scheduled boolean, If the item is scheduled.

scheduledDays array, Days of the week, based on numeric value starting at 0 for Sunday. Mutually exclusive with scheduledDaysOfMonth

scheduledHours array, Hours of the day it is scheduled on.

scheduledMinutes array, Minutes of the day it is scheduled on.

scheduledRunsPerHour integer, Deprecated in favor of scheduled minutes.

scheduledDaysOfMonth array, Days of the month it is scheduled on, mutually exclusive with scheduledDays.

parentId

integer, The ID of the parent job that will trigger this model.

timeZone

string, The time zone of this model.

lastRun

list, A list containing the following elements:

id integer,

state string,

createdAt string, The time that the run was queued.

startedAt string, The time that the run started.

finishedAt string, The time that the run completed.

error string, The error message for this run, if present.

user

list, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

createdAt

string, The time the model was created.

updatedAt

string, The time the model was updated.

currentBuildState

string, The status of the current model build. One of "succeeded", "failed", "queued", or "running,"or "idle", if no build has been attempted.

currentBuildException

string, Exception message, if applicable, of the current model build.

builds

array, An array containing the following fields:

id integer, The ID of the model build.

name string, The name of the model build.

createdAt string, The time the model build was created.

description string, A description of the model build.

rootMeanSquaredError number, A key metric for continuous models. Nil for other model types.

rSquaredError number, A key metric for continuous models. Nil for other model types.

rocAuc number, A key metric for binary, multinomial, and ordinal models. Nil for other model types.

predictions

array, An array containing the following fields:

id integer, The ID of the model to which to apply the prediction.

tableName string, The qualified name of the table on which to apply the predictive model.

primaryKey array, The primary key or composite keys of the table being predicted.

limitingSQL string, A SQL WHERE clause used to scope the rows to be predicted.

outputTable string, The qualified name of the table to be created which will contain the model's predictions.

state string, The status of the prediction. One of: "succeeded", "failed", "queued", or "running,"or "idle", if no build has been attempted.

lastOutputLocation

string, The output JSON for the last build.

archived

string, The archival status of the requested item(s).