Update some attributes of this Notebook

Update some attributes of this Notebook

notebooks_patch(id, name = NULL, language = NULL, description = NULL,
  file_id = NULL, requirements_file_id = NULL, requirements = NULL,
  docker_image_name = NULL, docker_image_tag = NULL,
  instance_type = NULL, memory = NULL, cpu = NULL,
  credentials = NULL, environment_variables = NULL,
  idle_timeout = NULL, git_repo_url = NULL, git_ref = NULL,
  git_path = NULL)

Arguments

id
integer required. The ID for this notebook.

name
string optional. The name of this notebook.

language
string optional. The kernel language of this notebook.

description
string optional. The description of this notebook.

file_id
string optional. The file ID for the S3 file containing the .ipynb file.

requirements_file_id
string optional. The file ID for the S3 file containing the requirements.txt file.

requirements
string optional. The requirements txt file.

docker_image_name
string optional. The name of the docker image to pull from DockerHub.

docker_image_tag
string optional. The tag of the docker image to pull from DockerHub (default: latest).

instance_type
string optional. The EC2 instance type to deploy to.

memory
integer optional. The amount of memory allocated to the notebook.

cpu
integer optional. The amount of cpu allocated to the the notebook.

credentials
array optional. A list of credential IDs to pass to the notebook.

environment_variables
list optional. Environment variables to be passed into the Notebook.

idle_timeout
integer optional. How long the notebook will stay alive without any kernel activity.

git_repo_url
string optional. The url of the git repository

git_ref
string optional. The git reference if git repo is specified

git_path
string optional. The path to the .ipynb file in the git repo that will be started up on notebook launch

Value

A list containing the following elements:

id

integer, The ID for this notebook.

name

string, The name of this notebook.

language

string, The kernel language of this notebook.

description

string, The description of this notebook.

notebookUrl

string, Time-limited URL to get the .ipynb file for this notebook.

notebookPreviewUrl

string, Time-limited URL to get the .htm preview file for this notebook.

requirementsUrl

string, Time-limited URL to get the requirements.txt file for this notebook.

fileId

string, The file ID for the S3 file containing the .ipynb file.

requirementsFileId

string, The file ID for the S3 file containing the requirements.txt file.

user

list, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

dockerImageName

string, The name of the docker image to pull from DockerHub.

dockerImageTag

string, The tag of the docker image to pull from DockerHub (default: latest).

instanceType

string, The EC2 instance type to deploy to.

memory

integer, The amount of memory allocated to the notebook.

cpu

integer, The amount of cpu allocated to the the notebook.

createdAt

string,

updatedAt

string,

mostRecentDeployment

list, A list containing the following elements:

deploymentId integer, The ID for this deployment.

userId integer, The ID of the owner.

host string, Domain of the deployment.

name string, Name of the deployment.

dockerImageName string, The name of the docker image to pull from DockerHub.

dockerImageTag string, The tag of the docker image to pull from DockerHub (default: latest).

displayUrl string, A signed URL for viewing the deployed item.

instanceType string, The EC2 instance type requested for the deployment.

memory integer, The memory allocated to the deployment.

cpu integer, The cpu allocated to the deployment.

state string, The state of the deployment.

stateMessage string, A detailed description of the state.

createdAt string,

updatedAt string,

published boolean,

notebookId integer, The ID of owning Notebook

credentials

array, A list of credential IDs to pass to the notebook.

environmentVariables

list, Environment variables to be passed into the Notebook.

idleTimeout

integer, How long the notebook will stay alive without any kernel activity.

gitRepoId

integer, The ID of the git repository.

gitRepoUrl

string, The url of the git repository

gitRef

string, The git reference if git repo is specified

gitPath

string, The path to the .ipynb file in the git repo that will be started up on notebook launch

archived

string, The archival status of the requested item(s).

hidden

boolean, The hidden status of the item.