List projects

List projects

projects_list(
  author = NULL,
  permission = NULL,
  hidden = NULL,
  archived = NULL,
  limit = NULL,
  page_num = NULL,
  order = NULL,
  order_dir = NULL
)

Arguments

author
string optional. If specified, return projects owned by this author. It accepts a comma-separated list of author ids.

permission
string optional. A permissions string, one of "read", "write", or "manage". Lists only projects for which the current user has that permission.

hidden
boolean optional. If specified to be true, returns hidden items. Defaults to false, returning non-hidden items.

archived
string optional. The archival status of the requested item(s).

limit
integer optional. Number of results to return. Defaults to 20. Maximum allowed is 1000.

page_num
integer optional. Page number of the results to return. Defaults to the first page, 1.

order
string optional. The field on which to order the result set. Defaults to updated_at. Must be one of: updated_at, name, created_at.

order_dir
string optional. Direction in which to sort, either asc (ascending) or desc (descending) defaulting to desc.

Value

An array containing the following fields:

id

integer, The ID for this project.

author

list, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

name

string, The name of this project.

description

string, A description of the project.

users

array, An array containing the following fields:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

autoShare

boolean,

createdAt

string,

updatedAt

string,

archived

string, The archival status of the requested item(s).