Get details about a query

Get details about a query

queries_get(id)

Arguments

id
integer required. The query ID.

Value

A list containing the following elements:

id

integer, The query ID.

database

integer, The database ID.

sql

string, The SQL to execute.

credential

integer, The credential ID.

resultRows

array, A preview of rows returned by the query.

resultColumns

array, A preview of columns returned by the query.

scriptId

integer, The ID of the script associated with this query.

exception

string, Exception returned from the query, null if the query was a success.

createdAt

string,

updatedAt

string,

finishedAt

string, The end time of the last run.

state

string, The state of the last run.

lastRunId

integer, The ID of the last run.

hidden

boolean, The hidden status of the object.

name

string, The name of the query.

author

object, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

startedAt

string, The start time of the last run.

reportId

integer, The ID of the report associated with this query.