Decommission an expired free trial

Decommission an expired free trial

registrations_post_decommission(id)

Arguments

id
integer required. The ID of the registration.

Value

A list containing the following elements:

id

integer, The ID of the registration.

firstName

string, This user's first name.

lastName

string, This user's last name.

username

string, This user's username.

email

string, This user's email address.

phone

string, This user's phone.

organizationName

string, The name of this user's organization.

comment

array, Array of comments for multi-select inputs.

status

string, The status of the registration (pending/confirmed/approved/rejected/provisioning/provisioned/notified/decommissioned).

ipAddress

string, The IP address used to create the registration.

createdAt

string, The date and time when the registration was created.

confirmedAt

string, The date and time when the registration was confirmed.

decommissionedAt

string, The date and time when the registration was decommissioned.

marketoLeadId

integer, The ID of the associated lead in Marketo, if it exists.

provisionJobId

integer, The ID of the provisioning job, if it has been created.

organizationId

integer, The ID of the free trial organization, if it has been created.

userId

integer, The ID of the free trial user, if it has been created.