List

List

reports_list(type = NULL, author = NULL, template_id = NULL,
  archived = NULL, limit = NULL, page_num = NULL, order = NULL,
  order_dir = NULL)

Arguments

type
string optional. If specified, return report of these types. It accepts a comma-separated list, possible values are 'tableau' or 'other'.

author
string optional. If specified, return reports from this author. It accepts a comma-separated list of author ids.

template_id
integer optional. If specified, return reports using the provided Template.

archived
string optional. The archival status of the requested object(s).

limit
integer optional. Number of results to return. Defaults to 20. Maximum allowed is 50.

page_num
integer optional. Page number of the results to return. Defaults to the first page, 1.

order
string optional. The field on which to order the result set. Defaults to updated_at. Must be one of: updated_at, name, created_at.

order_dir
string optional. Direction in which to sort, either asc (ascending) or desc (descending) defaulting to desc.

Value

An array containing the following fields:

id

integer, The ID of this report.

name

string, The name of the report.

user

object, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

createdAt

string,

updatedAt

string,

projects

array, An array containing the following fields:

id integer, The ID for the project.

name string, The name of the project.

state

string, The status of the report's last run.

finishedAt

string, The time that the report's last run finished.

vizUpdatedAt

string, The time that the report's visualization was last updated.

script

object, A list containing the following elements:

id integer, The ID for the script.

name string, The name of the script.

sql string, The raw SQL query for the script.

jobPath

string, The link to details of the job that backs this report.

tableauId

integer,

templateId

integer, The ID of the template used for this report.

authThumbnailUrl

string, URL for a thumbnail of the report.

lastRun

object, A list containing the following elements:

id integer,

state string,

createdAt string, The time that the run was queued.

startedAt string, The time that the run started.

finishedAt string, The time that the run completed.

error string, The error message for this run, if present.

archived

string, The archival status of the requested object(s).