Update a report

Update a report

reports_patch(
  id,
  name = NULL,
  script_id = NULL,
  code_body = NULL,
  config = NULL,
  app_state = NULL,
  provide_api_key = NULL,
  template_id = NULL,
  use_viewers_tableau_username = NULL,
  description = NULL
)

Arguments

id

integer required. The ID of the report to modify.

name

string optional. The name of the report.

script_id

integer optional. The ID of the job (a script or a query) used to create this report.

code_body

string optional. The code for the report visualization.

config

string optional.

app_state

list optional. The application state blob for this report.

provide_api_key

boolean optional. Allow the report to provide an API key to front-end code.

template_id

integer optional. The ID of the template used for this report. If null is passed, no template will back this report. Changes to the backing template will reset the report appState.

use_viewers_tableau_username

boolean optional. Apply user level filtering on Tableau reports.

description

string optional. The user-defined description of the report.

Value

A list containing the following elements:

id

integer, The ID of this report.

name

string, The name of the report.

user

list, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

createdAt

string, The creation time for this report.

updatedAt

string, The last updated at time for this report.

type

string, The type of the report. One of: ReportTypes::HTML, ReportTypes::Tableau, ReportTypes::ShinyApp, ReportTypes::SQL

description

string, The user-defined description of the report.

myPermissionLevel

string, Your permission level on the object. One of "read", "write", or "manage".

archived

string, The archival status of the requested item(s).

projects

array, An array containing the following fields:

id integer, The ID for the project.

name string, The name of the project.

state

string, The status of the report's last run.

finishedAt

string, The time that the report's last run finished.

vizUpdatedAt

string, The time that the report's visualization was last updated.

script

list, A list containing the following elements:

id integer, The ID for the script.

name string, The name of the script.

sql string, The raw SQL query for the script, if applicable.

jobPath

string, The link to details of the job that backs this report.

tableauId

integer,

templateId

integer, The ID of the template used for this report.

authThumbnailUrl

string, URL for a thumbnail of the report.

lastRun

list, A list containing the following elements:

id integer,

state string,

createdAt string, The time that the run was queued.

startedAt string, The time that the run started.

finishedAt string, The time that the run completed.

error string, The error message for this run, if present.

hidden

boolean, The hidden status of the item.

authDataUrl

string, DEPRECATED: For legacy reports

authCodeUrl

string, Link to code to render in the report.

config

string, Any configuration metadata for this report.

validOutputFile

boolean, Whether the job (a script or a query) that backs the report currently has a valid output file.

provideAPIKey

boolean, Whether the report requests an API Key from the report viewer.

apiKey

string, A Civis API key that can be used by this report.

apiKeyId

integer, The ID of the API key. Can be used for auditing API use by this report.

appState

list, Any application state blob for this report.

useViewersTableauUsername

boolean, Apply user level filtering on Tableau reports.