Update a SQL report

Update a SQL report

reports_patch_sql(
  id,
  query_id = NULL,
  name = NULL,
  config = NULL,
  description = NULL
)

Arguments

id

integer required. The ID of this report.

query_id

integer optional. The ID of the query used to create this report.

name

string optional. The name of the report.

config

string optional. The configuration of the report visualization.

description

string optional. The user-defined description of the report.

Value

A list containing the following elements:

id

integer, The ID of this report.

name

string, The name of the report.

user

list, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

createdAt

string, The creation time for this report.

updatedAt

string, The last updated at time for this report.

type

string, The type of the report. One of: ReportTypes::HTML, ReportTypes::Tableau, ReportTypes::ShinyApp, ReportTypes::SQL

description

string, The user-defined description of the report.

myPermissionLevel

string, Your permission level on the object. One of "read", "write", or "manage".

archived

string, The archival status of the requested item(s).

config

string, The configuration of the report visualization.

query

list, A list containing the following elements:

id integer, The query ID.

database integer, The database ID.

sql string, The SQL to execute.

credential integer, The credential ID.

resultRows array, A preview of rows returned by the query.

resultColumns array, A preview of columns returned by the query.

error string, The error message for this run, if present.

startedAt string, The start time of the last run.

finishedAt string, The end time of the last run.

state string, The state of the last run. One of queued, running, succeeded, failed, and cancelled.

runningAs list . A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.