Create a report

Create a report

results_post(script_id = NULL, name = NULL, code_body = NULL,
  app_state = NULL, provide_api_key = NULL, template_id = NULL,
  hidden = NULL)

Arguments

script_id
integer optional. The ID of the job (a script or a query) used to create this report.

name
string optional. The name of the report.

code_body
string optional. The code for the report visualization.

app_state
list optional. Any application state blob for this report.

provide_api_key
boolean optional. Allow the report to provide an API key to front-end code.

template_id
integer optional. The ID of the template used for this report.

hidden
boolean optional. The hidden status of the item.

Value

A list containing the following elements:

id

integer, The ID of this report.

name

string, The name of the report.

user

list, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

createdAt

string,

updatedAt

string,

projects

array, An array containing the following fields:

id integer, The ID for the project.

name string, The name of the project.

state

string, The status of the report's last run.

finishedAt

string, The time that the report's last run finished.

vizUpdatedAt

string, The time that the report's visualization was last updated.

script

list, A list containing the following elements:

id integer, The ID for the script.

name string, The name of the script.

sql string, The raw SQL query for the script.

jobPath

string, The link to details of the job that backs this report.

tableauId

integer,

type

string,

templateId

integer, The ID of the template used for this report.

authThumbnailUrl

string, URL for a thumbnail of the report.

lastRun

list, A list containing the following elements:

id integer,

state string,

createdAt string, The time that the run was queued.

startedAt string, The time that the run started.

finishedAt string, The time that the run completed.

error string, The error message for this run, if present.

archived

string, The archival status of the requested item(s).

hidden

boolean, The hidden status of the item.

authDataUrl

string,

authCodeUrl

string,

config

string, Any configuration metadata for this report.

validOutputFile

boolean, Whether the job (a script or a query) that backs the report currently has a valid output file.

provideAPIKey

boolean, Whether the report requests an API Key from the report viewer.

apiKey

string, A Civis API key that can be used by this report.

apiKeyId

integer, The ID of the API key. Can be used for auditing API use by this report.

appState

list, Any application state blob for this report.

useViewersTableauUsername

boolean, Apply user level filtering on Tableau reports.