List Scripts

List Scripts

scripts_list(
  type = NULL,
  category = NULL,
  author = NULL,
  status = NULL,
  hidden = NULL,
  archived = NULL,
  limit = NULL,
  page_num = NULL,
  order = NULL,
  order_dir = NULL
)

Arguments

type

string optional. If specified, return items of these types. The valid types are sql, python3, javascript, r, containers, and dbt.

category

string optional. A job category for filtering scripts. Must be one of script, import, export, and enhancement.

author

string optional. If specified, return items from any of these authors. It accepts a comma-separated list of user IDs.

status

string optional. If specified, returns items with one of these statuses. It accepts a comma-separated list, possible values are 'running', 'failed', 'succeeded', 'idle', 'scheduled'.

hidden

boolean optional. If specified to be true, returns hidden items. Defaults to false, returning non-hidden items.

archived

string optional. The archival status of the requested item(s).

limit

integer optional. Number of results to return. Defaults to 20. Maximum allowed is 50.

page_num

integer optional. Page number of the results to return. Defaults to the first page, 1.

order

string optional. The field on which to order the result set. Defaults to updated_at. Must be one of: updated_at, name, created_at, last_run.updated_at.

order_dir

string optional. Direction in which to sort, either asc (ascending) or desc (descending) defaulting to desc.

Value

An array containing the following fields:

id

integer, The ID for the script.

name

string, The name of the script.

type

string, The type of the script (e.g SQL, Container, Python, R, JavaScript, dbt)

createdAt

string, The time this script was created.

updatedAt

string, The time the script was last updated.

author

list, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

state

string, The status of the script's last run.

finishedAt

string, The time that the script's last run finished.

projects

array, An array containing the following fields:

id integer, The ID for the project.

name string, The name of the project.

parentId

integer, The ID of the parent job that will trigger this script

isTemplate

boolean, Whether others scripts use this one as a template.

fromTemplateId

integer, The ID of the template this script uses, if any.

links

list, A list containing the following elements:

details string, The details link to get more information about the script.

runs string, The runs link to get the run information list for this script.

timeZone

string, The time zone of this script.

lastRun

list, A list containing the following elements:

id integer,

state string,

createdAt string, The time that the run was queued.

startedAt string, The time that the run started.

finishedAt string, The time that the run completed.

error string, The error message for this run, if present.

archived

string, The archival status of the requested item(s).

templateScriptId

integer, The ID of the template script, if any.