List Custom Scripts

List Custom Scripts

scripts_list_custom(from_template_id = NULL, author = NULL, status = NULL,
  hidden = NULL, archived = NULL, limit = NULL, page_num = NULL,
  order = NULL, order_dir = NULL)

Arguments

from_template_id
string optional. If specified, return scripts based on the template with this ID. Specify multiple IDs as a comma-separated list.

author
string optional. If specified, return items from this author. Must use user IDs. A comma separated list of IDs is also accepted to return items from multiple authors.

status
string optional. If specified, returns items with one of these statuses. It accepts a comma-separated list, possible values are 'running', 'failed', 'succeeded', 'idle', 'scheduled'.

hidden
boolean optional. If specified to be true, returns hidden items. Defaults to false, returning non-hidden items.

archived
string optional. The archival status of the requested item(s).

limit
integer optional. Number of results to return. Defaults to its maximum of 50.

page_num
integer optional. Page number of the results to return. Defaults to the first page, 1.

order
string optional. The field on which to order the result set. Defaults to updated_at. Must be one of: updated_at, name, created_at.

order_dir
string optional. Direction in which to sort, either asc (ascending) or desc (descending) defaulting to asc.

Value

An array containing the following fields:

id

integer, The ID for the script.

name

string, The name of the script.

type

string, The type of the script (e.g Custom)

createdAt

string, The time this script was created.

updatedAt

string, The time the script was last updated.

author

list, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

state

string, The status of the script's last run.

finishedAt

string, The time that the script's last run finished.

projects

array, An array containing the following fields:

id integer, The ID for the project.

name string, The name of the project.

parentId

integer, The ID of the parent job that will trigger this script

fromTemplateId

integer, The ID of the template script.

timeZone

string, The time zone of this script.

lastRun

list, A list containing the following elements:

id integer,

state string,

createdAt string, The time that the run was queued.

startedAt string, The time that the run started.

finishedAt string, The time that the run completed.

error string, The error message for this run, if present.

archived

string, The archival status of the requested item(s).