Update a container

Update a container

scripts_patch_containers(id, name = NULL, parent_id = NULL,
  user_context = NULL, params = NULL, arguments = NULL, schedule = NULL,
  notifications = NULL, required_resources = NULL, repo_http_uri = NULL,
  repo_ref = NULL, remote_host_credential_id = NULL,
  git_credential_id = NULL, docker_command = NULL,
  docker_image_name = NULL, docker_image_tag = NULL,
  cancel_timeout = NULL, time_zone = NULL, target_project_id = NULL)

Arguments

id
integer required. The ID for the script.

name
string optional. The name of the container.

parent_id
integer optional. The ID of the parent job that will trigger this script

user_context
string optional. "runner" or "author", who to execute the script as when run as a template.

params
array optional. An array containing the following fields:

name string, The variable's name as used within your code.

label string, The label to present to users when asking them for the value.

description string, A short sentence or fragment describing this parameter to the end user.

type string, The type of parameter. Valid options: string, integer, float, bool, file, database, credential_aws, credential_redshift, or credential_custom

required boolean, Whether this param is required.

value string, The value you would like to set this param to. Setting this value makes this parameter a fixed param.

default string, If an argument for this parameter is not defined, it will use this default value. Use true, True, t, y, yes, or 1 for true bool's or false, False, f, n, no, or 0 for false bool's. Cannot be used for parameters that are required or a credential type.

arguments
object optional. Dictionary of name/value pairs to use to run this script. Only settable if this script has defined params.

schedule
object optional. A list containing the following elements:

scheduled boolean, If the object is scheduled

scheduledDays array, Day based on numeric value starting at 0 for Sunday

scheduledHours array, Hours of the day it is scheduled on

scheduledMinutes array, Minutes of the day it is scheduled on

scheduledRunsPerHour integer, Alternative to scheduled minutes, number of times to run per hour

notifications
object optional. A list containing the following elements:

urls array, URLs to receive a POST request at job completion

successEmailSubject string, Custom subject line for success e-mail.

successEmailBody string, Custom body text for success e-mail, written in Markdown.

successEmailAddresses array, Addresses to notify by e-mail when the job completes successfully.

failureEmailAddresses array, Addresses to notify by e-mail when the job fails.

stallWarningMinutes integer, Stall warning emails will be sent after this amount of minutes.

successOn boolean, If success email notifications are on

failureOn boolean, If failure email notifications are on

required_resources
object optional. A list containing the following elements:

cpu integer, The number of CPU shares to allocate for the container. Each core has 1024 shares.

memory integer, The amount of RAM to allocate for the container (in MiB).

diskSpace number, The amount of disk space, in GB, to allocate for the container. This space will be used to hold the git repo configured for the container and anything your container writes to /tmp or /data. Fractional values (e.g. 0.25) are supported.

repo_http_uri
string optional. The location of a github repo to clone into the container, e.g. github.com/my-user/my-repo.git.

repo_ref
string optional. The tag or branch of the github repo to clone into the container.

remote_host_credential_id
integer optional. The id of the database credentials to pass into the environment of the container.

git_credential_id
integer optional. The id of the git credential to be used when checking out the specified git repo. If not supplied, the first git credential you've submitted will be used. Unnecessary if no git repo is specified or the git repo is public.

docker_command
string optional. The command to run on the container. Will be run via sh as: ["sh", "-c", dockerCommand]

docker_image_name
string optional. The name of the docker image to pull from DockerHub.

docker_image_tag
string optional. The tag of the docker image to pull from DockerHub (default: latest).

cancel_timeout
integer optional. The amount of time (in seconds) to wait before forcibly terminating the script. When the script is cancelled, it is first sent a TERM signal. If the script is still running after the timeout, it is sent a KILL signal. Defaults to 0.

time_zone
string optional. The time zone of this script.

target_project_id
integer optional. Target project to which script outputs will be added.

Value

A list containing the following elements:

id

integer, The ID for the script.

name

string, The name of the container.

type

string, The type of the script (e.g Container)

createdAt

string, The time this script was created.

updatedAt

string, The time the script was last updated.

author

object, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

state

string, The status of the script's last run.

finishedAt

string, The time that the script's last run finished.

projects

array, An array containing the following fields:

id integer, The ID for the project.

name string, The name of the project.

parentId

integer, The ID of the parent job that will trigger this script

userContext

string, "runner" or "author", who to execute the script as when run as a template.

params

array, An array containing the following fields:

name string, The variable's name as used within your code.

label string, The label to present to users when asking them for the value.

description string, A short sentence or fragment describing this parameter to the end user.

type string, The type of parameter. Valid options: string, integer, float, bool, file, database, credential_aws, credential_redshift, or credential_custom

required boolean, Whether this param is required.

value string, The value you would like to set this param to. Setting this value makes this parameter a fixed param.

default string, If an argument for this parameter is not defined, it will use this default value. Use true, True, t, y, yes, or 1 for true bool's or false, False, f, n, no, or 0 for false bool's. Cannot be used for parameters that are required or a credential type.

arguments

object, Dictionary of name/value pairs to use to run this script. Only settable if this script has defined params.

isTemplate

boolean, Whether others scripts use this one as a template.

templateDependentsCount

integer, How many other scripts use this one as a template.

publishedAsTemplateId

integer, The ID of the template that this script is backing.

fromTemplateId

integer, The ID of the template script.

templateScriptName

string, The name of the template script.

links

object, A list containing the following elements:

details string, The details link to get more information about the script.

runs string, The runs link to get the run information list for this script.

schedule

object, A list containing the following elements:

scheduled boolean, If the object is scheduled

scheduledDays array, Day based on numeric value starting at 0 for Sunday

scheduledHours array, Hours of the day it is scheduled on

scheduledMinutes array, Minutes of the day it is scheduled on

scheduledRunsPerHour integer, Alternative to scheduled minutes, number of times to run per hour

notifications

object, A list containing the following elements:

urls array, URLs to receive a POST request at job completion

successEmailSubject string, Custom subject line for success e-mail.

successEmailBody string, Custom body text for success e-mail, written in Markdown.

successEmailAddresses array, Addresses to notify by e-mail when the job completes successfully.

failureEmailAddresses array, Addresses to notify by e-mail when the job fails.

stallWarningMinutes integer, Stall warning emails will be sent after this amount of minutes.

successOn boolean, If success email notifications are on

failureOn boolean, If failure email notifications are on

runningAs

object, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

requiredResources

object, A list containing the following elements:

cpu integer, The number of CPU shares to allocate for the container. Each core has 1024 shares.

memory integer, The amount of RAM to allocate for the container (in MiB).

diskSpace number, The amount of disk space, in GB, to allocate for the container. This space will be used to hold the git repo configured for the container and anything your container writes to /tmp or /data. Fractional values (e.g. 0.25) are supported.

repoHttpUri

string, The location of a github repo to clone into the container, e.g. github.com/my-user/my-repo.git.

repoRef

string, The tag or branch of the github repo to clone into the container.

remoteHostCredentialId

integer, The id of the database credentials to pass into the environment of the container.

gitCredentialId

integer, The id of the git credential to be used when checking out the specified git repo. If not supplied, the first git credential you've submitted will be used. Unnecessary if no git repo is specified or the git repo is public.

dockerCommand

string, The command to run on the container. Will be run via sh as: ["sh", "-c", dockerCommand]

dockerImageName

string, The name of the docker image to pull from DockerHub.

dockerImageTag

string, The tag of the docker image to pull from DockerHub (default: latest).

cancelTimeout

integer, The amount of time (in seconds) to wait before forcibly terminating the script. When the script is cancelled, it is first sent a TERM signal. If the script is still running after the timeout, it is sent a KILL signal. Defaults to 0.

lastRun

object, A list containing the following elements:

id integer,

state string,

createdAt string, The time that the run was queued.

startedAt string, The time that the run started.

finishedAt string, The time that the run completed.

error string, The error message for this run, if present.

timeZone

string, The time zone of this script.

hidden

boolean, The hidden status of the object. Setting this to true hides it from most API endpoints. The object can still be queried directly by ID

archived

string, The archival status of the requested object(s).

targetProjectId

integer, Target project to which script outputs will be added.