Update an attached git file

Update an attached git file

scripts_patch_r_git(
  id,
  git_ref = NULL,
  git_branch = NULL,
  git_path = NULL,
  git_repo_url = NULL,
  git_ref_type = NULL,
  pull_from_git = NULL
)

Arguments

id

integer required. The ID of the item.

git_ref

string optional. A git reference specifying an unambiguous version of the file. Can be a branch name, or the full or shortened SHA of a commit.

git_branch

string optional. The git branch that the file is on.

git_path

string optional. The path of the file in the repository.

git_repo_url

string optional. The URL of the git repository (e.g., https://github.com/organization/repo_name.git).

git_ref_type

string optional. Specifies if the file is versioned by branch or tag.

pull_from_git

boolean optional. Automatically pull latest commit from git. Only works for scripts.

Value

A list containing the following elements:

gitRef

string, A git reference specifying an unambiguous version of the file. Can be a branch name, tag or the full or shortened SHA of a commit.

gitBranch

string, The git branch that the file is on.

gitPath

string, The path of the file in the repository.

gitRepo

list, A list containing the following elements:

id integer, The ID for this git repository.

repoUrl string, The URL for this git repository.

createdAt string,

updatedAt string,

gitRefType

string, Specifies if the file is versioned by branch or tag.

pullFromGit

boolean, Automatically pull latest commit from git. Only works for scripts and workflows (assuming you have the feature enabled)