Create a Custom Script

Create a Custom Script

scripts_post_custom(
  from_template_id,
  name = NULL,
  parent_id = NULL,
  arguments = NULL,
  remote_host_id = NULL,
  credential_id = NULL,
  schedule = NULL,
  notifications = NULL,
  time_zone = NULL,
  hidden = NULL,
  target_project_id = NULL,
  required_resources = NULL,
  partition_label = NULL,
  running_as_id = NULL
)

Arguments

from_template_id

integer required. The ID of the template script.

name

string optional. The name of the script.

parent_id

integer optional. The ID of the parent job that will trigger this script

arguments

list optional. Parameter-value pairs to use when running this script. Only settable if this script has defined parameters.

remote_host_id

integer optional. The remote host ID that this script will connect to.

credential_id

integer optional. The credential that this script will use.

schedule

list optional. A list containing the following elements:

scheduled boolean, If the item is scheduled.

scheduledDays array, Days of the week, based on numeric value starting at 0 for Sunday. Mutually exclusive with scheduledDaysOfMonth

scheduledHours array, Hours of the day it is scheduled on.

scheduledMinutes array, Minutes of the day it is scheduled on.

scheduledRunsPerHour integer, Deprecated in favor of scheduled minutes.

scheduledDaysOfMonth array, Days of the month it is scheduled on, mutually exclusive with scheduledDays.

notifications

list optional. A list containing the following elements:

urls array, URLs to receive a POST request at job completion

successEmailSubject string, Custom subject line for success e-mail.

successEmailBody string, Custom body text for success e-mail, written in Markdown.

successEmailAddresses array, Addresses to notify by e-mail when the job completes successfully.

successEmailFromName string, Name from which success emails are sent; defaults to "Civis."

successEmailReplyTo string, Address for replies to success emails; defaults to the author of the job.

failureEmailAddresses array, Addresses to notify by e-mail when the job fails.

stallWarningMinutes integer, Stall warning emails will be sent after this amount of minutes.

successOn boolean, If success email notifications are on. Defaults to user's preferences.

failureOn boolean, If failure email notifications are on. Defaults to user's preferences.

time_zone

string optional. The time zone of this script.

hidden

boolean optional. The hidden status of the item.

target_project_id

integer optional. Target project to which script outputs will be added.

required_resources

list optional. A list containing the following elements:

cpu integer, The number of CPU shares to allocate for the container. Each core has 1000 shares.

memory integer, The amount of RAM to allocate for the container (in MB).

diskSpace number, The amount of disk space, in GB, to allocate for the container. This space will be used to hold the git repo configured for the container and anything your container writes to /tmp or /data. Fractional values (e.g. 0.25) are supported.

partition_label

string optional. The partition label used to run this object. Only applicable for jobs using Docker.

running_as_id

integer optional. The ID of the runner of this script.

Value

A list containing the following elements:

id

integer, The ID for the script.

fromTemplateAliases

array, An array containing the following fields:

id integer, The id of the Alias object.

objectId integer, The id of the object

alias string, The alias of the object

name

string, The name of the script.

type

string, The type of the script (e.g Custom)

backingScriptType

string, The type of the script backing this template (e.g Python)

createdAt

string, The time this script was created.

updatedAt

string, The time the script was last updated.

author

list, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

state

string, The status of the script's last run.

finishedAt

string, The time that the script's last run finished.

category

string,

projects

array, An array containing the following fields:

id integer, The ID for the project.

name string, The name of the project.

parentId

integer, The ID of the parent job that will trigger this script

params

array, An array containing the following fields:

name string, The variable's name as used within your code.

label string, The label to present to users when asking them for the value.

description string, A short sentence or fragment describing this parameter to the end user.

type string, The type of parameter. Valid options: string, multi_line_string, integer, float, bool, file, table, database, credential_aws, credential_redshift, or credential_custom

required boolean, Whether this param is required.

value string, The value you would like to set this param to. Setting this value makes this parameter a fixed param.

default string, If an argument for this parameter is not defined, it will use this default value. Use true, True, t, y, yes, or 1 for true bool's or false, False, f, n, no, or 0 for false bool's. Cannot be used for parameters that are required or a credential type.

allowedValues array, The possible values this parameter can take, effectively making this an enumerable parameter. Allowed values is an array of hashes of the following format: `{label: 'Import', 'value': 'import'}`

arguments

list, Parameter-value pairs to use when running this script. Only settable if this script has defined parameters.

isTemplate

boolean, Whether others scripts use this one as a template.

publishedAsTemplateId

integer, The ID of the template that this script is backing.

fromTemplateId

integer, The ID of the template script.

uiReportUrl

integer, The url of the custom HTML.

uiReportId

integer, The id of the report with the custom HTML.

uiReportProvideAPIKey

boolean, Whether the ui report requests an API Key from the report viewer.

templateScriptName

string, The name of the template script.

templateNote

string, The template's note.

remoteHostId

integer, The remote host ID that this script will connect to.

credentialId

integer, The credential that this script will use.

codePreview

string, The code that this script will run with arguments inserted.

schedule

list, A list containing the following elements:

scheduled boolean, If the item is scheduled.

scheduledDays array, Days of the week, based on numeric value starting at 0 for Sunday. Mutually exclusive with scheduledDaysOfMonth

scheduledHours array, Hours of the day it is scheduled on.

scheduledMinutes array, Minutes of the day it is scheduled on.

scheduledRunsPerHour integer, Deprecated in favor of scheduled minutes.

scheduledDaysOfMonth array, Days of the month it is scheduled on, mutually exclusive with scheduledDays.

notifications

list, A list containing the following elements:

urls array, URLs to receive a POST request at job completion

successEmailSubject string, Custom subject line for success e-mail.

successEmailBody string, Custom body text for success e-mail, written in Markdown.

successEmailAddresses array, Addresses to notify by e-mail when the job completes successfully.

successEmailFromName string, Name from which success emails are sent; defaults to "Civis."

successEmailReplyTo string, Address for replies to success emails; defaults to the author of the job.

failureEmailAddresses array, Addresses to notify by e-mail when the job fails.

stallWarningMinutes integer, Stall warning emails will be sent after this amount of minutes.

successOn boolean, If success email notifications are on. Defaults to user's preferences.

failureOn boolean, If failure email notifications are on. Defaults to user's preferences.

runningAs

list, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

timeZone

string, The time zone of this script.

lastRun

list, A list containing the following elements:

id integer,

state string,

createdAt string, The time that the run was queued.

startedAt string, The time that the run started.

finishedAt string, The time that the run completed.

error string, The error message for this run, if present.

myPermissionLevel

string, Your permission level on the object. One of "read", "write", or "manage".

hidden

boolean, The hidden status of the item.

archived

string, The archival status of the requested item(s).

targetProjectId

integer, Target project to which script outputs will be added.

lastSuccessfulRun

list, A list containing the following elements:

id integer,

state string,

createdAt string, The time that the run was queued.

startedAt string, The time that the run started.

finishedAt string, The time that the run completed.

error string, The error message for this run, if present.

requiredResources

list, A list containing the following elements:

cpu integer, The number of CPU shares to allocate for the container. Each core has 1000 shares.

memory integer, The amount of RAM to allocate for the container (in MB).

diskSpace number, The amount of disk space, in GB, to allocate for the container. This space will be used to hold the git repo configured for the container and anything your container writes to /tmp or /data. Fractional values (e.g. 0.25) are supported.

partitionLabel

string, The partition label used to run this object. Only applicable for jobs using Docker.

runningAsId

integer, The ID of the runner of this script.