Perform a search

Perform a search

search_list(query = NULL, type = NULL, offset = NULL, order = NULL,
  owner = NULL, limit = NULL, archived = NULL)

Arguments

query
string optional. The search query.

type
string optional. The type for the search. It accepts a comma-separated list. Valid arguments are listed on the "GET /search/types" endpoint.

offset
integer optional. The offset for the search results.

order
string optional. The field on which to order the result set.

owner
string optional. The owner for the search.

limit
integer optional. Defaults to 10. Maximum allowed is 50.

archived
string optional. If specified, return only results with the chosen archived status; either 'true', 'false', or 'all'. Defaults to 'false'.

Value

An array containing the following fields:

totalResults

integer, The number of items matching the search query.

aggregations

list, Aggregations by owner and type for the search results.

results

array, An array containing the following fields:

score number, The relevance score from the search request.

type string, The type of the item.

id integer, The ID of the item.

name string, The name of the item.

typeName string, The verbose name of the type.

updatedAt string, The time the item was last updated.

owner string, The owner of the item.

useCount integer, The use count of the item, if the item is a template.