Search queries that are not hidden

Search queries that are not hidden

search_list_queries(
  search_string = NULL,
  database_id = NULL,
  credential_id = NULL,
  author_id = NULL,
  archived = NULL,
  state = NULL,
  started_before = NULL,
  started_after = NULL,
  limit = NULL,
  page_num = NULL,
  order = NULL,
  order_dir = NULL
)

Arguments

search_string

string optional. Space delimited search terms for searching queries by their SQL. Supports wild card characters "?" for any single character, and "*" for zero or more characters.

database_id

integer optional. The database ID.

credential_id

integer optional. The credential ID.

author_id

integer optional. The author of the query.

archived

boolean optional. The archival status of the requested item(s). Defaults to false.

state

array optional. The state of the last run. One or more of queued, running, succeeded, failed, and cancelled.

started_before

string optional. An upper bound for the start date of the last run.

started_after

string optional. A lower bound for the start date of the last run.

limit

integer optional. Number of results to return. Defaults to 10. Maximum allowed is 50.

page_num

integer optional. Page number of the results to return. Defaults to the first page, 1.

order

string optional. The field on which to order the result set. Defaults to last_run_started_at. Must be one of: last_run_started_at.

order_dir

string optional. Direction in which to sort, either asc (ascending) or desc (descending) defaulting to desc.

Value

An array containing the following fields:

id

integer, The query ID.

database

integer, The database ID.

credential

integer, The credential ID.

sql

string, The SQL executed by the query.

authorId

integer, The author of the query.

archived

boolean, The archival status of the requested item(s).

createdAt

string,

updatedAt

string,

lastRun

list, A list containing the following elements:

id integer,

state string, The state of the run. One of queued, running, succeeded, failed, and cancelled.

startedAt string, The time that the run started.

finishedAt string, The time that the run completed.

error string, The error message for this run, if present.