List Services

List Services

services_list(
  hidden = NULL,
  archived = NULL,
  author = NULL,
  status = NULL,
  limit = NULL,
  page_num = NULL,
  order = NULL,
  order_dir = NULL
)

Arguments

hidden

boolean optional. If specified to be true, returns hidden items. Defaults to false, returning non-hidden items.

archived

string optional. The archival status of the requested item(s).

author

string optional. If specified, return items from any of these authors. It accepts a comma-separated list of user IDs.

status

string optional. If specified, returns Services with one of these statuses. It accepts a comma-separated list, possible values are 'running', 'idle'.

limit

integer optional. Number of results to return. Defaults to 20. Maximum allowed is 50.

page_num

integer optional. Page number of the results to return. Defaults to the first page, 1.

order

string optional. The field on which to order the result set. Defaults to updated_at. Must be one of: updated_at, name, created_at.

order_dir

string optional. Direction in which to sort, either asc (ascending) or desc (descending) defaulting to desc.

Value

An array containing the following fields:

id

integer, The ID for this Service.

name

string, The name of this Service.

description

string, The description of this Service.

user

list, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

type

string, The type of this Service

createdAt

string,

updatedAt

string,

gitRepoUrl

string, The url for the git repo where the Service code lives.

gitRepoRef

string, The git reference to use when pulling code from the repo.

gitPathDir

string, The path to the Service code within the git repo. If unspecified, the root directory will be used.

currentDeployment

list, A list containing the following elements:

deploymentId integer, The ID for this deployment.

userId integer, The ID of the owner.

host string, Domain of the deployment.

name string, Name of the deployment.

dockerImageName string, The name of the docker image to pull from DockerHub.

dockerImageTag string, The tag of the docker image to pull from DockerHub (default: latest).

instanceType string, The EC2 instance type requested for the deployment.

memory integer, The memory allocated to the deployment, in MB.

cpu integer, The cpu allocated to the deployment, in millicores.

state string, The state of the deployment.

stateMessage string, A detailed description of the state.

maxMemoryUsage number, If the deployment has finished, the maximum amount of memory used during the deployment, in MB.

maxCpuUsage number, If the deployment has finished, the maximum amount of cpu used during the deployment, in millicores.

createdAt string,

updatedAt string,

serviceId integer, The ID of owning Service

archived

string, The archival status of the requested item(s).