Standardize addresses in a table

Standardize addresses in a table

tables_post_enhancements_cass_ncoa(source_table_id, perform_ncoa = NULL,
  ncoa_credential_id = NULL, output_level = NULL)

Arguments

source_table_id
integer required. The ID of the table to be enhanced.

perform_ncoa
boolean optional. Whether to update addresses for records matching the National Change of Address (NCOA) database.

ncoa_credential_id
integer optional. Credential to use when performing NCOA updates. Required if 'performNcoa' is true.

output_level
string optional. The set of fields persisted by a CASS or NCOA enhancement.For CASS enhancements, one of 'cass' or 'all.'For NCOA enhancements, one of 'cass', 'ncoa' , 'coalesced' or 'all'.By default, all fields will be returned.

Value

A list containing the following elements:

id

integer, The ID of the enhancement.

sourceTableId

integer, The ID of the table that was enhanced.

state

string, The state of the enhancement, one of 'queued' 'running' 'succeeded' 'failed' or 'cancelled'.

enhancedTableSchema

string, The schema name of the table created by the enhancement.

enhancedTableName

string, The name of the table created by the enhancement.

performNcoa

boolean, Whether to update addresses for records matching the National Change of Address (NCOA) database.

ncoaCredentialId

integer, Credential to use when performing NCOA updates. Required if 'performNcoa' is true.

outputLevel

string, The set of fields persisted by a CASS or NCOA enhancement.For CASS enhancements, one of 'cass' or 'all.'For NCOA enhancements, one of 'cass', 'ncoa' , 'coalesced' or 'all'.By default, all fields will be returned.