Request a refresh for column and table statistics

Request a refresh for column and table statistics

tables_post_refresh(id)

Arguments

id
integer required.

Value

A list containing the following elements:

id

integer, The ID of the table.

databaseId

integer, The ID of the database.

schema

string, The name of the schema containing the table.

name

string, Name of the table.

description

string, The description of the table, as specified by the table owner

isView

boolean, True if this table represents a view. False if it represents a regular table.

rowCount

integer, The number of rows in the table.

columnCount

integer, The number of columns in the table.

sizeMb

number, The size of the table in megabytes.

owner

string, The database username of the table's owner.

distkey

string, The column used as the Amazon Redshift distkey.

sortkeys

string, The column used as the Amazon Redshift sortkey.

refreshStatus

string, How up-to-date the table's statistics on row counts, null counts, distinct counts, and values distributions are. One of: refreshing, stale, or current.

lastRefresh

string, The time of the last statistics refresh.

refreshId

string, The ID of the most recent statistics refresh.

lastRun

list, A list containing the following elements:

id integer,

state string,

createdAt string, The time that the run was queued.

startedAt string, The time that the run started.

finishedAt string, The time that the run completed.

error string, The error message for this run, if present.

ontologyMapping

list, The ontology-key to column-name mapping. See /ontology for the list of valid ontology keys.

columns

array, An array containing the following fields:

name string, Name of the column.

sqlType string, SQL type of the column.

sampleValues array, A sample of values from the column.

encoding string, The compression encoding for this columnSee: http://docs.aws.amazon.com/redshift/latest/dg/c_Compression_encodings.html

description string, The description of the column, as specified by the table owner

order integer, Relative position of the column in the table.

minValue string, Smallest value in the column.

maxValue string, Largest value in the column.

avgValue number, Average value of the column, where applicable.

stddev number, Stddev of the column, where applicable.

valueDistributionPercent object, A mapping between each value in the column and the percentage of rows with that value.Only present for tables with fewer than approximately 25,000,000 rows and for columns with fewer than twenty distinct values.

coverageCount integer, Number of non-null values in the column.

nullCount integer, Number of null values in the column.

possibleDependentVariableTypes array, Possible dependent variable types the column may be used to model. Null if it may not be used as a dependent variable.

useableAsIndependentVariable boolean, Whether the column may be used as an independent variable to train a model.

useableAsPrimaryKey boolean, Whether the column may be used as an primary key to identify table rows.

valueDistribution object, An object mapping distinct values in the column to the number of times they appear in the column

distinctCount integer, Number of distinct values in the column.

joins

array, An array containing the following fields:

id integer,

leftTableId integer,

leftIdentifier string,

rightTableId integer,

rightIdentifier string,

on string,

leftJoin boolean,

createdAt string,

updatedAt string,

multipartKey

array,

enhancements

array, An array containing the following fields:

type string,

createdAt string,

updatedAt string,

joinId integer,

viewDef

string,

tableDef

string,

outgoingTableMatches

array, An array containing the following fields:

sourceTableId integer, Source table

targetType string, Target type

targetId integer, Target ID

target object,

job object,