Review a template for security vulnerability and correctness (admin-only)

Review a template for security vulnerability and correctness (admin-only)

templates_post_reports_review(id, status)

Arguments

id
integer required. The ID of the item.

status
boolean required. Whether this item has been reviewed.

Value

A list containing the following elements:

id

integer,

name

string, The name of the template.

category

string, The category of this report template. Can be left blank. Acceptable values are: dataset-viz

createdAt

string,

updatedAt

string,

useCount

integer, The number of uses of this template.

archived

boolean, Whether the template has been archived.

author

list, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

techReviewed

boolean, Whether this template has been audited by Civis for security vulnerability and correctness.

authCodeUrl

string, A URL to the template's stored code body.

provideAPIKey

boolean, Whether reports based on this template request an API Key from the report viewer.

hidden

boolean, The hidden status of the item.