Show info about a user

Show info about a user

users_get(id)

Arguments

id

integer required. The ID of this user.

Value

A list containing the following elements:

id

integer, The ID of this user.

user

string, The username of this user.

name

string, The name of this user.

email

string, The email of this user.

active

boolean, Whether this user account is active or deactivated.

primaryGroupId

integer, The ID of the primary group of this user.

groups

array, An array containing the following fields:

id integer, The ID of this group.

name string, The name of this group.

slug string, The slug of this group.

organizationId integer, The ID of the organization associated with this group.

organizationName string, The name of the organization associated with this group.

city

string, The city of this user.

state

string, The state of this user.

timeZone

string, The time zone of this user.

initials

string, The initials of this user.

department

string, The department of this user.

title

string, The title of this user.

githubUsername

string, The GitHub username of this user.

prefersSmsOtp

boolean, The preference for phone authorization of this user

vpnEnabled

boolean, The availability of vpn for this user.

ssoDisabled

boolean, The availability of SSO for this user.

otpRequiredForLogin

boolean, The two factor authentication requirement for this user.

exemptFromOrgSmsOtpDisabled

boolean, Whether the user has SMS OTP enabled on an individual level. This field does not matter if the org does not have SMS OTP disabled.

smsOtpAllowed

boolean, Whether the user is allowed to receive two factor authentication codes via SMS.

robot

boolean, Whether the user is a robot.

phone

string, The phone number of this user.

organizationSlug

string, The slug of the organization the user belongs to.

organizationSSODisableCapable

boolean, The user's organization's ability to disable sso for their users.

organizationLoginType

string, The user's organization's login type.

organizationSmsOtpDisabled

boolean, Whether the user's organization has SMS OTP disabled.

myPermissionLevel

string, Your permission level on the object. One of "read", "write", or "manage".

createdAt

string, The date and time when the user was created.

updatedAt

string, The date and time when the user was last updated.

lastSeenAt

string, The date and time when the user last visited Platform.

suspended

boolean, Whether the user is suspended due to inactivity.

createdById

integer, The ID of the user who created this user.

lastUpdatedById

integer, The ID of the user who last updated this user.

unconfirmedEmail

string, The new email address awaiting confirmation from the user.

accountStatus

string, Account status of this user. One of: "Active", "Deactivated", "Suspended", "Unsuspended"