List users

List users

users_list(
  feature_flag = NULL,
  account_status = NULL,
  query = NULL,
  group_id = NULL,
  group_ids = NULL,
  organization_id = NULL,
  exclude_groups = NULL,
  limit = NULL,
  page_num = NULL,
  order = NULL,
  order_dir = NULL
)

Arguments

feature_flag

string optional. Return users that have a feature flag enabled.

account_status

string optional. The account status by which to filter users. May be one of "active", "inactive", or "all". Defaults to active.

query

string optional. Return users who match the given query, based on name, user, email, and id.

group_id

integer optional. The ID of the group by which to filter users. Cannot be present if group_ids is.

group_ids

array optional. The IDs of the groups by which to filter users. Cannot be present if group_id is.

organization_id

integer optional. The ID of the organization by which to filter users.

exclude_groups

boolean optional. Whether or to exclude users' groups. Default: false.

limit

integer optional. Number of results to return. Defaults to 20. Maximum allowed is 10000.

page_num

integer optional. Page number of the results to return. Defaults to the first page, 1.

order

string optional. The field on which to order the result set. Defaults to name. Must be one of: name, user.

order_dir

string optional. Direction in which to sort, either asc (ascending) or desc (descending) defaulting to asc.

Value

An array containing the following fields:

id

integer, The ID of this user.

user

string, The username of this user.

name

string, The name of this user.

email

string, The email of this user.

active

boolean, Whether this user account is active or deactivated.

primaryGroupId

integer, The ID of the primary group of this user.

groups

array, An array containing the following fields:

id integer, The ID of this group.

name string, The name of this group.

slug string, The slug of this group.

organizationId integer, The ID of the organization associated with this group.

organizationName string, The name of the organization associated with this group.

createdAt

string, The date and time when the user was created.

currentSignInAt

string, The date and time when the user's current session began.

updatedAt

string, The date and time when the user was last updated.

lastSeenAt

string, The date and time when the user last visited Platform.

suspended

boolean, Whether the user is suspended due to inactivity.

createdById

integer, The ID of the user who created this user.

lastUpdatedById

integer, The ID of the user who last updated this user.