Show API keys belonging to the specified user

Show API keys belonging to the specified user

users_list_api_keys(
  id,
  limit = NULL,
  page_num = NULL,
  order = NULL,
  order_dir = NULL
)

Arguments

id

string required. The ID of the user or 'me'.

limit

integer optional. Number of results to return. Defaults to its maximum of 50.

page_num

integer optional. Page number of the results to return. Defaults to the first page, 1.

order

string optional. The field on which to order the result set. Defaults to id. Must be one of: id.

order_dir

string optional. Direction in which to sort, either asc (ascending) or desc (descending) defaulting to desc.

Value

An array containing the following fields:

id

integer, The ID of the API key.

name

string, The name of the API key.

expiresAt

string, The date and time when the key expired.

createdAt

string, The date and time when the key was created.

revokedAt

string, The date and time when the key was revoked.

lastUsedAt

string, The date and time when the key was last used.

scopes

array, The scopes which the key is permissioned on.

useCount

integer, The number of times the key has been used.

expired

boolean, True if the key has expired.

active

boolean, True if the key has neither expired nor been revoked.

constraintCount

integer, The number of constraints on the created key