Update info about the logged-in user

Update info about the logged-in user

users_patch_me(preferences = NULL, last_checked_announcements = NULL)

Arguments

preferences
object optional. A list containing the following elements:

appIndexOrderField string, Order field for the apps index pages.

appIndexOrderDir string, Oder direction for the apps index pages.

resultIndexOrderField string, Order field for the results index page.

resultIndexOrderDir string, Order direction for the results index page.

resultIndexTypeFilter string, Type filter for the results index page.

resultIndexAuthorFilter string, Author filter for the results index page.

resultIndexArchivedFilter string, Archived filter for the results index page.

importIndexOrderField string, Order field for the imports index page.

importIndexOrderDir string, Order direction for the imports index page.

importIndexTypeFilter string, Type filter for the imports index page.

importIndexAuthorFilter string, Author filter for the imports index page.

importIndexDestFilter string, Destination filter for the imports index page.

importIndexStatusFilter string, Status filter for the imports index page.

importIndexArchivedFilter string, Archived filter for the imports index page.

exportIndexOrderField string, Order field for the exports index page.

exportIndexOrderDir string, Order direction for the exports index page.

exportIndexTypeFilter string, Type filter for the exports index page.

exportIndexAuthorFilter string, Author filter for the exports index page.

exportIndexStatusFilter string, Status filter for the exports index page.

modelIndexOrderField string, Order field for the models index page.

modelIndexOrderDir string, Order direction for the models index page.

modelIndexAuthorFilter string, Author filter for the models index page.

modelIndexStatusFilter string, Status filter for the models index page.

modelIndexArchivedFilter string, Archived filter for the models index page.

modelIndexThumbnailView string, Thumbnail view for the models index page.

scriptIndexOrderField string, Order field for the scripts index page.

scriptIndexOrderDir string, Order direction for the scripts index page.

scriptIndexTypeFilter string, Type filter for the scripts index page.

scriptIndexAuthorFilter string, Author filter for the scripts index page.

scriptIndexStatusFilter string, Status filter for the scripts index page.

scriptIndexArchivedFilter string, Archived filter for the scripts index page.

projectIndexOrderField string, Order field for the projects index page.

projectIndexOrderDir string, Order direction for the projects index page.

projectIndexAuthorFilter string, Author filter for the projects index page.

projectIndexArchivedFilter string, Archived filter for the projects index page.

reportIndexThumbnailView string, Thumbnail view for the reports index page.

projectDetailOrderField string, Order field for projects detail pages.

projectDetailOrderDir string, Order direction for projects detail pages.

projectDetailAuthorFilter string, Author filter for projects detail pages.

projectDetailTypeFilter string, Type filter for projects detail pages.

projectDetailArchivedFilter string, Archived filter for the projects detail pages.

enhancementIndexOrderField string, Order field for the enhancements index page.

enhancementIndexOrderDir string, Order direction for the enhancements index page.

enhancementIndexAuthorFilter string, Author filter for the enhancements index page.

enhancementIndexArchivedFilter string, Archived filter for the enhancements index page.

preferredServerId integer, ID of preferred server.

civisExploreSkipIntro boolean, Whether the user is shown steps for each exploration.

registrationIndexOrderField string, Order field for the registrations index page.

registrationIndexOrderDir string, Order direction for the registrations index page.

registrationIndexStatusFilter string, Status filter for the registrations index page.

upgradeRequested string, Whether a free trial upgrade has been requested.

welcomeOrderField string, Order direction for the welcome page.

welcomeOrderDir string, Order direction for the welcome page.

welcomeAuthorFilter string, Status filter for the welcome page.

welcomeStatusFilter string, Status filter for the welcome page.

welcomeArchivedFilter string, Status filter for the welcome page.

dataPaneWidth string, Width of the data pane when expanded.

dataPaneCollapsed string, Whether the data pane is collapsed.

notebookOrderField string, Order field for the notebooks page.

notebookOrderDir string, Order direction for the notebooks page.

notebookAuthorFilter string, Author filter for the notebooks page.

notebookArchivedFilter string, Archived filter for the notebooks page.

notebookStatusFilter string, Status filter for the notebooks page.

workflowIndexOrderField string, Order field for the workflows page.

workflowIndexOrderDir string, Order direction for the workflows page.

workflowIndexAuthorFilter string, Author filter for the workflows page.

serviceOrderField string, Order field for the services page.

serviceOrderDir string, Order direction for the services page.

serviceAuthorFilter string, Author filter for the services page.

serviceArchivedFilter string, Archived filter for the services page.

last_checked_announcements
string optional. The date and time at which the user last checked their announcements.

Value

A list containing the following elements:

id

integer, The ID of this user.

name

string, This user's name.

email

string, This user's email address.

username

string, This user's username.

initials

string, This user's initials.

lastCheckedAnnouncements

string, The date and time at which the user last checked their announcements.

featureFlags

object, The feature flag settings for this user.

roles

array, The roles this user has, listed by slug.

preferences

object, This user's preferences.

customBranding

string, The branding of Platform for this user.

groups

array, An array containing the following fields:

id integer, The ID of this group.

name string, The name of this group.

organizationId integer, The organization associated with this group.

organizationName

string, The name of the organization the user belongs to.

createdAt

string, The date and time when the user was created.

signInCount

integer, The number of times the user has signed in.