Create a new user (must be an admin or client user admin)

Create a new user (must be an admin or client user admin)

users_post(
  name,
  email,
  primary_group_id,
  user,
  active = NULL,
  city = NULL,
  state = NULL,
  time_zone = NULL,
  initials = NULL,
  department = NULL,
  title = NULL,
  prefers_sms_otp = NULL,
  group_ids = NULL,
  vpn_enabled = NULL,
  sso_disabled = NULL,
  otp_required_for_login = NULL,
  exempt_from_org_sms_otp_disabled = NULL,
  robot = NULL,
  send_email = NULL
)

Arguments

name
string required. The name of this user.

email
string required. The email of this user.

primary_group_id
integer required. The ID of the primary group of this user.

user
string required. The username of this user.

active
boolean optional. The account status of this user.

city
string optional. The city of this user.

state
string optional. The state of this user.

time_zone
string optional. The time zone of this user.

initials
string optional. The initials of this user.

department
string optional. The department of this user.

title
string optional. The title of this user.

prefers_sms_otp
boolean optional. The preference for phone authorization of this user

group_ids
array optional. An array of ids of all the groups this user is in.

vpn_enabled
boolean optional. The availability of vpn for this user.

sso_disabled
boolean optional. The availability of SSO for this user.

otp_required_for_login
boolean optional. The two factor authentication requirement for this user.

exempt_from_org_sms_otp_disabled
boolean optional. Whether the user has SMS OTP enabled on an individual level. This field does not matter if the org does not have SMS OTP disabled.

robot
boolean optional. Whether the user is a robot.

send_email
boolean optional. Whether the user will receive a welcome email.

Value

A list containing the following elements:

id

integer, The ID of this user.

user

string, The username of this user.

name

string, The name of this user.

email

string, The email of this user.

active

boolean, The account status of this user.

primaryGroupId

integer, The ID of the primary group of this user.

groups

array, An array containing the following fields:

id integer, The ID of this group.

name string, The name of this group.

organizationId integer, The organization associated with this group.

city

string, The city of this user.

state

string, The state of this user.

timeZone

string, The time zone of this user.

initials

string, The initials of this user.

department

string, The department of this user.

title

string, The title of this user.

githubUsername

string, The GitHub username of this user.

prefersSmsOtp

boolean, The preference for phone authorization of this user

vpnEnabled

boolean, The availability of vpn for this user.

ssoDisabled

boolean, The availability of SSO for this user.

otpRequiredForLogin

boolean, The two factor authentication requirement for this user.

exemptFromOrgSmsOtpDisabled

boolean, Whether the user has SMS OTP enabled on an individual level. This field does not matter if the org does not have SMS OTP disabled.

smsOtpAllowed

boolean, Whether the user is allowed to receive two factor authentication codes via SMS.

robot

boolean, Whether the user is a robot.

phone

string, The phone number of this user.

organizationSlug

string, The slug of the organization the user belongs to.

organizationSSODisableCapable

boolean, The user's organization's ability to disable sso for their users.

organizationLoginType

string, The user's organization's login type.

organizationSmsOtpDisabled

boolean, Whether the user's organization has SMS OTP disabled.