Create a Workflow

Create a Workflow

workflows_post(
  name,
  description = NULL,
  from_job_chain = NULL,
  definition = NULL,
  schedule = NULL,
  allow_concurrent_executions = NULL,
  time_zone = NULL,
  notifications = NULL,
  hidden = NULL
)

Arguments

name

string required. The name of this workflow.

description

string optional. A description of the workflow.

from_job_chain

integer optional. If specified, create a workflow from the job chain this job is in, and inherit the schedule from the root of the chain.

definition

string optional. The definition of the workflow in YAML format. Must not be specified if `fromJobChain` is specified.

schedule

list optional. A list containing the following elements:

scheduled boolean, If the item is scheduled.

scheduledDays array, Days of the week, based on numeric value starting at 0 for Sunday. Mutually exclusive with scheduledDaysOfMonth

scheduledHours array, Hours of the day it is scheduled on.

scheduledMinutes array, Minutes of the day it is scheduled on.

scheduledRunsPerHour integer, Deprecated in favor of scheduled minutes.

scheduledDaysOfMonth array, Days of the month it is scheduled on, mutually exclusive with scheduledDays.

allow_concurrent_executions

boolean optional. Whether the workflow can execute when already running.

time_zone

string optional. The time zone of this workflow.

notifications

list optional. A list containing the following elements:

urls array, URLs to receive a POST request at job completion

successEmailSubject string, Custom subject line for success e-mail.

successEmailBody string, Custom body text for success e-mail, written in Markdown.

successEmailAddresses array, Addresses to notify by e-mail when the job completes successfully.

failureEmailAddresses array, Addresses to notify by e-mail when the job fails.

stallWarningMinutes integer, Stall warning emails will be sent after this amount of minutes.

successOn boolean, If success email notifications are on

failureOn boolean, If failure email notifications are on

hidden

boolean optional. The hidden status of the item.

Value

A list containing the following elements:

id

integer, The ID for this workflow.

name

string, The name of this workflow.

description

string, A description of the workflow.

definition

string, The definition of the workflow in YAML format. Must not be specified if `fromJobChain` is specified.

valid

boolean, The validity of the workflow definition.

validationErrors

string, The errors encountered when validating the workflow definition.

fileId

string, The file id for the s3 file containing the workflow configuration.

user

list, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

state

string, The state of the workflow. State is "running" if any execution is running, otherwise reflects most recent execution state.

lastExecutionId

integer, The ID of the most recent execution of this workflow, if any.

schedule

list, A list containing the following elements:

scheduled boolean, If the item is scheduled.

scheduledDays array, Days of the week, based on numeric value starting at 0 for Sunday. Mutually exclusive with scheduledDaysOfMonth

scheduledHours array, Hours of the day it is scheduled on.

scheduledMinutes array, Minutes of the day it is scheduled on.

scheduledRunsPerHour integer, Deprecated in favor of scheduled minutes.

scheduledDaysOfMonth array, Days of the month it is scheduled on, mutually exclusive with scheduledDays.

allowConcurrentExecutions

boolean, Whether the workflow can execute when already running.

timeZone

string, The time zone of this workflow.

nextExecutionAt

string, The time of the next scheduled execution.

notifications

list, A list containing the following elements:

urls array, URLs to receive a POST request at job completion

successEmailSubject string, Custom subject line for success e-mail.

successEmailBody string, Custom body text for success e-mail, written in Markdown.

successEmailAddresses array, Addresses to notify by e-mail when the job completes successfully.

failureEmailAddresses array, Addresses to notify by e-mail when the job fails.

stallWarningMinutes integer, Stall warning emails will be sent after this amount of minutes.

successOn boolean, If success email notifications are on

failureOn boolean, If failure email notifications are on

archived

string, The archival status of the requested item(s).

hidden

boolean, The hidden status of the item.

myPermissionLevel

string, Your permission level on the object. One of "read", "write", or "manage".

createdAt

string,

updatedAt

string,