Evaluate an expression in Civis Platform

Evaluate an expression in Civis Platform

CivisFuture(expr = NULL, envir = parent.frame(), substitute = FALSE,
  globals = TRUE, packages = NULL, lazy = FALSE, local = TRUE,
  gc = FALSE, earlySignal = FALSE, label = NULL,
  required_resources = list(cpu = 1024, memory = 2048, diskSpace = 4),
  docker_image_name = "civisanalytics/datascience-r",
  docker_image_tag = "2.3.0", ...)

Arguments

expr
An R expression.

envir
The environment in which the evaluation is done (or inherits from if local is TRUE).

substitute
If TRUE, argument expr is substitute():ed, otherwise not.

globals
(optional) a named list of global objects needed in order for the future to be resolved correctly.

packages
(optional) a character vector specifying packages to be attached in the R environment evaluating the future.

lazy
If FALSE (default), the future is resolved eagerly (starting immediately), otherwise not.

local
If TRUE, the expression is evaluated such that all assignments are done to local temporary environment, otherwise the assignments are done to the global environment of the R process evaluating the future.

gc
If TRUE, the garbage collector run (in the process that evaluated the future) after the value of the future is collected. Some types of futures ignore this argument.

earlySignal
Specified whether conditions should be signaled as soon as possible or not.

label
An optional character string label attached to the future.

required_resources
resources, see scripts_post_containers

docker_image_name
the image for the container script.

docker_image_tag
the tag for the Docker image.

...
arguments to scripts_post_containers

Value

A CivisFuture inheriting from Future that evaluates expr on the given container.