Evaluate an expression in Civis Platform

Evaluate an expression in Civis Platform

CivisFuture(
  expr = NULL,
  envir = parent.frame(),
  substitute = FALSE,
  globals = TRUE,
  packages = NULL,
  lazy = FALSE,
  local = TRUE,
  gc = FALSE,
  earlySignal = FALSE,
  label = NULL,
  required_resources = list(cpu = 1024, memory = 2048, diskSpace = 4),
  docker_image_name = "civisanalytics/datascience-r",
  docker_image_tag = "3",
  ...
)

# S3 method for CivisFuture
run(future, ...)

# S3 method for CivisFuture
result(future, ...)

# S3 method for CivisFuture
cancel(future, ...)

# S3 method for CivisFuture
resolved(future, ...)

# S3 method for CivisFuture
fetch_logs(object, ...)

Arguments

expr
An R expression.

envir
The environment from where global objects should be identified.

substitute
If TRUE, argument expr is substitute():ed, otherwise not.

globals
(optional) a logical, a character vector, or a named list to control how globals are handled. For details, see section 'Globals used by future expressions' in the help for future().

packages
(optional) a character vector specifying packages to be attached in the R environment evaluating the future.

lazy
If FALSE (default), the future is resolved eagerly (starting immediately), otherwise not.

local
If TRUE, the expression is evaluated such that all assignments are done to local temporary environment, otherwise the assignments are done to the global environment of the R process evaluating the future.

gc
If TRUE, the garbage collector run (in the process that evaluated the future) only after the value of the future is collected. Exactly when the values are collected may depend on various factors such as number of free workers and whether earlySignal is TRUE (more frequently) or FALSE (less frequently). Some types of futures ignore this argument.

earlySignal
Specified whether conditions should be signaled as soon as possible or not.

label
An optional character string label attached to the future.

required_resources
resources, see scripts_post_containers

docker_image_name
the image for the container script.

docker_image_tag
the tag for the Docker image.

...
arguments to scripts_post_containers

future
CivisFuture object.

object
CivisFuture

Value

A CivisFuture inheriting from Future that evaluates expr on the given container.

Methods (by generic)

run: Run a CivisFuture

result: Return the value of a CivisFuture

cancel: Cancel a CivisFuture

resolved: Check if a CivisFutre has resolved

fetch_logs: Fetch logs from a CivisFuture