Initate an import of a tabular file into the platform

Initate an import of a tabular file into the platform

imports_post_files(schema, name, remote_host_id, credential_id,
  max_errors = NULL, existing_table_rows = NULL, distkey = NULL,
  sortkey1 = NULL, sortkey2 = NULL, column_delimiter = NULL,
  first_row_is_header = NULL, multipart = NULL, hidden = NULL)

Arguments

schema
string required. The schema of the destination table.

name
string required. The name of the destination table.

remote_host_id
integer required. The id of the destination database host.

credential_id
integer required. The id of the credentials to be used when performing the database import.

max_errors
integer optional. The maximum number of rows with errors to remove from the import before failing.

existing_table_rows
string optional. The behaviour if a table with the requested name already exists. One of "fail", "truncate", "append", or "drop".Defaults to "fail".

distkey
string optional. The column to use as the distkey for the table.

sortkey1
string optional. The column to use as the sort key for the table.

sortkey2
string optional. The second column in a compound sortkey for the table.

column_delimiter
string optional. The column delimiter of the file. If column_delimiter is null or omitted, it will be auto-detected. Valid arguments are "comma", "tab", and "pipe".

first_row_is_header
boolean optional. A boolean value indicating whether or not the first row is a header row. If first_row_is_header is null or omitted, it will be auto-detected.

multipart
boolean optional. If true, the upload URI will require a `multipart/form-data` POST request. Defaults to false.

hidden
boolean optional. The hidden status of the object. Setting this to true hides it from most API endpoints. The object can still be queried directly by ID

Value

A list containing the following elements:

id

integer, The id of the import.

uploadUri

string, The URI which may be used to upload a tabular file for import. You must use this URI to upload the file you wish imported and then inform the Civis API when your upload is complete using the URI given by the runUri field of this reponse.

runUri

string, The URI to POST to once the file upload is complete. After uploading the file using the URI given in the uploadUri attribute of the reponse, POST to this URI to initiate the import of your uploaded file into the platform.

uploadFields

object, If multipart was set to true, these fields should be included in the multipart upload.