Create a CSV Import

Create a CSV Import

imports_post_files_csv(source, destination, first_row_is_header,
  column_delimiter = NULL, escaped = NULL, compression = NULL,
  existing_table_rows = NULL, max_errors = NULL,
  table_columns = NULL, execution = NULL,
  redshift_destination_options = NULL, hidden = NULL)

Arguments

source
list required. A list containing the following elements:

fileIds array, The file ID(s) to import, if importing Civis file(s).

storagePaths object,

destination
list required. A list containing the following elements:

schema string, The destination schema name.

table string, The destination table name.

remoteHostId integer, The ID of the destination database host.

credentialId integer, The ID of the credentials for the destination database.

primaryKeys array, A list of the primary key column(s) of the destination table.If the destination table does not exist, and the import mode is "upsert", this field is required.

lastModifiedKeys array, A list of the columns indicating a record has been updated.If the destination table does not exist, and the import mode is "upsert", this field is required.

first_row_is_header
boolean required. A boolean value indicating whether or not the first row of the source file is a header row.

column_delimiter
string optional. The column delimiter for the file. Valid arguments are "comma", "tab", and "pipe". Defaults to "comma".

escaped
boolean optional. A boolean value indicating whether or not the source file has quotes escaped with a backslash.Defaults to false.

compression
string optional. The type of compression of the source file. Valid arguments are "gzip" and "none". Defaults to "none".

existing_table_rows
string optional. The behavior if a destination table with the requested name already exists. One of "fail", "truncate", "append", "drop", or "upsert".Defaults to "fail".

max_errors
integer optional. The maximum number of rows with errors to ignore before failing. This option is not supported for Postgres databases.

table_columns
array optional. An array containing the following fields:

name string, The column name.

sqlType string, The SQL type of the column.

execution
string optional. In upsert mode, controls the movement of data in upsert mode. If set to "delayed", the data will be moved after a brief delay. If set to "immediate", the data will be moved immediately. In non-upsert modes, controls the speed at which detailed column stats appear in the data catalogue. Defaults to "delayed", to accomodate concurrent upserts to the same table and speedier non-upsert imports.

redshift_destination_options
list optional. A list containing the following elements:

distkey string, Distkey for this table in Redshift

sortkeys array, Sortkeys for this table in Redshift. Please provide a maximum of two.

hidden
boolean optional. The hidden status of the item.

Value

A list containing the following elements:

id

integer, The ID for the import.

source

list, A list containing the following elements:

fileIds array, The file ID(s) to import, if importing Civis file(s).

storagePaths object,

destination

list, A list containing the following elements:

schema string, The destination schema name.

table string, The destination table name.

remoteHostId integer, The ID of the destination database host.

credentialId integer, The ID of the credentials for the destination database.

primaryKeys array, A list of the primary key column(s) of the destination table.If the destination table does not exist, and the import mode is "upsert", this field is required.

lastModifiedKeys array, A list of the columns indicating a record has been updated.If the destination table does not exist, and the import mode is "upsert", this field is required.

firstRowIsHeader

boolean, A boolean value indicating whether or not the first row of the source file is a header row.

columnDelimiter

string, The column delimiter for the file. Valid arguments are "comma", "tab", and "pipe". Defaults to "comma".

escaped

boolean, A boolean value indicating whether or not the source file has quotes escaped with a backslash.Defaults to false.

compression

string, The type of compression of the source file. Valid arguments are "gzip" and "none". Defaults to "none".

existingTableRows

string, The behavior if a destination table with the requested name already exists. One of "fail", "truncate", "append", "drop", or "upsert".Defaults to "fail".

maxErrors

integer, The maximum number of rows with errors to ignore before failing. This option is not supported for Postgres databases.

tableColumns

array, An array containing the following fields:

name string, The column name.

sqlType string, The SQL type of the column.

execution

string, In upsert mode, controls the movement of data in upsert mode. If set to "delayed", the data will be moved after a brief delay. If set to "immediate", the data will be moved immediately. In non-upsert modes, controls the speed at which detailed column stats appear in the data catalogue. Defaults to "delayed", to accomodate concurrent upserts to the same table and speedier non-upsert imports.

redshiftDestinationOptions

list, A list containing the following elements:

distkey string, Distkey for this table in Redshift

sortkeys array, Sortkeys for this table in Redshift. Please provide a maximum of two.

hidden

boolean, The hidden status of the item.