Publish an R Markdown file to Platform Reports

Publish an R Markdown file to Platform Reports

publish_rmd(rmd_file, report_id = NULL, report_name = NULL,
  provide_api_key = NULL, project_id = NULL, ...)

Arguments

rmd_file
string, R Markdown file (.Rmd)

report_id
integer, ID of an existing report, if provided, the contents of the report will be replaced. If report_id is NULL, a new report will be created.

report_name
string, Title of report in platform

provide_api_key
bool, Set to true to include API key in report.

project_id
bool, Project_id that the report should be added to.

...
additional parameters to send to rmarkdown::render. Note: A temporary file will be used for output_file if output_file is not explicity set and input will be overwritten with rmd_file.

Note

rmarkdown::render depends on a recent version of pandoc. pandoc is distributed with RStudio and thus, publish_rmd will work in an RStudio environment. Outside of RStudio pandoc may be installed or knitr::knit2html can be used to convert an R Markdown document into html. The html can then be published to Civis with publish_html.

See also

'publish_html' to publish html to Civis

Examples

# NOT RUN {
# Publish a standard report
publish_rmd("my_beautiful_doc.Rmd")

# Publish a parameterized R Markdown document
# See: http://rmarkdown.rstudio.com/developer_parameterized_reports.html
params <- list("region" = "east", start = as.Date("2015-02-01"))
publish_rmd("my_parameterized_doc.Rmd", params=params)
# }