Clone this Service

Clone this Service

services_post_clone(id)

Arguments

id
integer required.

Value

A list containing the following elements:

id

integer, The ID for this Service.

name

string, The name of this Service.

description

string, The description of this Service.

user

list, A list containing the following elements:

id integer, The ID of this user.

name string, This user's name.

username string, This user's username.

initials string, This user's initials.

online boolean, Whether this user is online.

type

string, The type of this Service

dockerImageName

string, The name of the docker image to pull from DockerHub.

dockerImageTag

string, The tag of the docker image to pull from DockerHub (default: latest).

schedule

list, A list containing the following elements:

runtimePlan string, Only affects the service when deployed. On Demand means that the service will be turned on when viewed and automatically turned off after periods of inactivity. Specific Times means the service will be on when scheduled. Always On means the deployed service will always be on.

recurrences array, List of day-hour combinations this item is scheduled for

timeZone

string,

replicas

integer, The number of Service replicas to deploy. When maxReplicas is set, this field defines the minimum number of replicas to deploy.

maxReplicas

integer, The maximum number of Service replicas to deploy. Defining this field enables autoscaling.

instanceType

string, The EC2 instance type to deploy to.

memory

integer, The amount of memory allocated to each replica of the Service.

cpu

integer, The amount of cpu allocated to each replica of the the Service.

createdAt

string,

updatedAt

string,

credentials

array, A list of credential IDs to pass to the Service.

apiKeyId

integer, API key id of user

gitRepoUrl

string, The url for the git repo where the Service code lives.

gitRepoRef

string, The git reference to use when pulling code from the repo.

gitPathDir

string, The path to the Service code within the git repo. If unspecified, the root directory will be used.

reportId

integer, The ID of the associated report.

currentDeployment

list, A list containing the following elements:

deploymentId integer, The ID for this deployment.

userId integer, The ID of the owner.

host string, Domain of the deployment.

name string, Name of the deployment.

dockerImageName string, The name of the docker image to pull from DockerHub.

dockerImageTag string, The tag of the docker image to pull from DockerHub (default: latest).

displayUrl string, A signed URL for viewing the deployed item.

instanceType string, The EC2 instance type requested for the deployment.

memory integer, The memory allocated to the deployment.

cpu integer, The cpu allocated to the deployment.

state string, The state of the deployment.

stateMessage string, A detailed description of the state.

createdAt string,

updatedAt string,

published boolean,

serviceId integer, The ID of owning Service

previewDeployment

list, A list containing the following elements:

deploymentId integer, The ID for this deployment.

userId integer, The ID of the owner.

host string, Domain of the deployment.

name string, Name of the deployment.

dockerImageName string, The name of the docker image to pull from DockerHub.

dockerImageTag string, The tag of the docker image to pull from DockerHub (default: latest).

displayUrl string, A signed URL for viewing the deployed item.

instanceType string, The EC2 instance type requested for the deployment.

memory integer, The memory allocated to the deployment.

cpu integer, The cpu allocated to the deployment.

state string, The state of the deployment.

stateMessage string, A detailed description of the state.

createdAt string,

updatedAt string,

published boolean,

serviceId integer, The ID of owning Service

currentUrl

string, The URL that the service is hosted at.

previewUrl

string, The URL that previews of the service are hosted at.

environmentVariables

list, Environment Variables to be passed into the Service.

notifications

list, A list containing the following elements:

failureEmailAddresses array, Addresses to notify by e-mail when the service fails.

failureOn boolean, If failure email notifications are on

archived

string, The archival status of the requested item(s).

hidden

boolean, The hidden status of the item.