List columns in the specified table

List columns in the specified table

tables_list_columns(
  id,
  name = NULL,
  limit = NULL,
  page_num = NULL,
  order = NULL,
  order_dir = NULL
)

Arguments

id
integer required.

name
string optional. Search for columns with the given name, within the specified table.

limit
integer optional. Number of results to return. Defaults to its maximum of 50.

page_num
integer optional. Page number of the results to return. Defaults to the first page, 1.

order
string optional. The field on which to order the result set. Defaults to name. Must be one of: name, order.

order_dir
string optional. Direction in which to sort, either asc (ascending) or desc (descending) defaulting to asc.

Value

An array containing the following fields:

name

string, Name of the column.

civisDataType

string, The generic data type of the column (ex. "string"). Since this is database-agnostic, it may be helpful when loading data to R/Python.

sqlType

string, The database-specific SQL type of the column (ex. "varchar(30)").

sampleValues

array, A sample of values from the column.

encoding

string, The compression encoding for this columnSee: http://docs.aws.amazon.com/redshift/latest/dg/c_Compression_encodings.html

description

string, The description of the column, as specified by the table owner

order

integer, Relative position of the column in the table.

minValue

string, Smallest value in the column.

maxValue

string, Largest value in the column.

avgValue

number, Average value of the column, where applicable.

stddev

number, Stddev of the column, where applicable.

valueDistributionPercent

list, A mapping between each value in the column and the percentage of rows with that value.Only present for tables with fewer than approximately 25,000,000 rows and for columns with fewer than twenty distinct values.

coverageCount

integer, Number of non-null values in the column.

nullCount

integer, Number of null values in the column.

possibleDependentVariableTypes

array, Possible dependent variable types the column may be used to model. Null if it may not be used as a dependent variable.

useableAsIndependentVariable

boolean, Whether the column may be used as an independent variable to train a model.

useableAsPrimaryKey

boolean, Whether the column may be used as an primary key to identify table rows.

valueDistribution

list, An object mapping distinct values in the column to the number of times they appear in the column

distinctCount

integer, Number of distinct values in the column.