Transfer a table from one location to another.

Transfer a table from one location to another.

transfer_table(source_db, dest_db, source_table, dest_table, job_name = NULL,
  source_credential_id = NULL, dest_credential_id = NULL, interval = NULL,
  verbose = FALSE, advanced_options = NULL)

Arguments

source_db
string or int, The name of the database where the source table is located. Optionally, could be the database ID.

dest_db
string or int, The name of the database where the table will be transfered. Optionally, could be the database ID.

source_table
string, Full name of the table to transfer, e.g., "schema.table".

dest_table
string, Full name of the table in the destination database, e.g., "schema.table".

job_name
string, optional, A name to give the job. If omitted, a random job name will be used.

source_credential_id
string or int, Optional credential ID for the source database. If NULL, the default credential will be used.

dest_credential_id
string or int, Optional credential ID for the source database. If NULL, the default credential will be used.

interval
Number of seconds to wait between checks for job completion. If NULL, the default exponential backoff from await will be used.

verbose
bool, Set to TRUE to print intermediate progress indicators.

advanced_options
A list of advanced options for the sync. See imports_post_syncs for details.

Value

A civis_api object.

See also

Other tables: get_table_id, refresh_table

Examples

# NOT RUN {
transfer_table(source_db='Cluster A', dest_db='Cluster B',
               source_table='schma.tbl', dest_table='schma.tbl')
# }