Show info about a user

Show info about a user

users_get(id)

Arguments

id
integer required. The ID of this user.

Value

A list containing the following elements:

id

integer, The ID of this user.

user

string, The username of this user.

name

string, The name of this user.

email

string, The email of this user.

active

boolean, The account status of this user.

primaryGroupId

integer, The ID of the primary group of this user.

groups

array, An array containing the following fields:

id integer, The ID of this group.

name string, The name of this group.

organizationId integer, The organization associated with this group.

city

string, The city of this user.

state

string, The state of this user.

timeZone

string, The time zone of this user.

initials

string, The initials of this user.

department

string, The department of this user.

title

string, The title of this user.

githubUsername

string, The GitHub username of this user.

prefersSmsOtp

boolean, The preference for phone authorization of this user

vpnEnabled

boolean, The availability of vpn for this user.

ssoDisabled

boolean, The availability of SSO for this user.

otpRequiredForLogin

boolean, The two factor authentication requirement for this user.

exemptFromOrgSmsOtpDisabled

boolean, Whether the user has SMS OTP enabled on an individual level. This field does not matter if the org does not have SMS OTP disabled.

smsOtpAllowed

boolean, Whether the user is allowed to receive two factor authentication codes via SMS.

robot

boolean, Whether the user is a robot.

phone

string, The phone number of this user.

organizationSlug

string, The slug of the organization the user belongs to.

organizationSSODisableCapable

boolean, The user's organization's ability to disable sso for their users.

organizationLoginType

string, The user's organization's login type.

organizationSmsOtpDisabled

boolean, Whether the user's organization has SMS OTP disabled.