Upload a local data frame or csv file to the Civis Platform (Redshift)

Uploads either a data frame or csv file to Redshift, based on the first argument.

A default database for all IO operations can be set using options(civis.default_db = "mydatabase").

write_civis(x, ...)

# S3 method for data.frame
write_civis(x, tablename, database = NULL,
  if_exists = "fail", distkey = NULL, sortkey1 = NULL, sortkey2 = NULL,
  max_errors = NULL, verbose = FALSE, ...)

# S3 method for character
write_civis(x, tablename, database = NULL,
  if_exists = "fail", distkey = NULL, sortkey1 = NULL, sortkey2 = NULL,
  max_errors = NULL, verbose = FALSE, ...)

Arguments

x
data frame or file path of csv to upload to platform.

...
Extra arguments to write.csv

tablename
string, Name of table and schema "schema.tablename"

database
string, Name of database where data frame is to be uploaded. If no database is specified, uses options(civis.default_db)

if_exists
string, optional, String indicating action to take if table already exists. Must be either "fail", "truncate" or "append". Defaults to "fail".

distkey
string, optional, Column name designating the distkey

sortkey1
string, optional, Column name designating the first sortkey

sortkey2
string, optional, Column name designating the second (compound) sortkey.

max_errors
int, optional, Maximum number of rows with errors to remove before failing.

verbose
bool, Set to TRUE to print intermediate progress indicators.

Methods (by class)

data.frame: Upload a data frame to Civis Platform (Redshift)

character: Upload a csv to Civis Platform (Redshift)

See also

refresh_table to update table meta-data.

Other io: download_civis, query_civis, read_civis, write_civis_file

Examples

# NOT RUN {
df <- read.csv(local_file)

# Create new table, fail if already exists
write_civis(df, "schema.my_table", "my_database")

# Create new table, append if already exists
write_civis(df, "schema.my_table", "my_database", if_exists="append")

# Create new table with defined diskey / sortkeys for speed!
write_civis(df, "schema.my_table", "my_database", distkey="id",
            sortkey1="added_date")

# Create new table directly from a saved csv
write_civis("my/file/path.csv", "schema.my_table", "my_database")

# }