Upload a R object or file to Civis Platform (Files endpoint)

Uploads a R object or file to the files endpoint on Civis Platform (Amazon S3). It returns the id of the file for use with read_civis or download_civis.

R objects are serialized with saveRDS, files are unserialized. Objects or files larger than 50mb are chunked and can be uploaded in parallel if a plan has been set. Files larger than 5TB cannot be uploaded.

write_civis_file(x, ...)

# S3 method for default
write_civis_file(x, name = "r-object.rds",
  expires_at = NULL, ...)

# S3 method for character
write_civis_file(x, name = x, expires_at = NULL,
  ...)

Arguments

x
R object or path of file to upload.

...
arguments passed to saveRDS.

name
string, Name of the file or object.

expires_at
string, The date and time the object will expire on in the format "YYYY-MM-DD HH:MM:SS". expires_at = NULL allows files to be kept indefinitely.

Value

The file id which can be used to later retrieve the file using read_civis.

Details

By default, R objects are serialized using saveRDS before uploading the object to the files endpoint. If given a filepath, the file is uploaded as-is.

Methods (by class)

default: Serialize R object to Civis Platform (Files endpoint).

character: Upload a text file to Civis Platform (Files endpoint).

See also

Examples

# NOT RUN {
data(iris)
file_id <- write_civis_file(iris)
read_civis(file_id)

file_id <- write_civis_file("path/to/my.csv")
read_civis(file_id, using = read.csv)
read_civis(file_id, using = readr::read_csv)

# Does not expire
file_id <- write_civis_file(iris, expires_at = NULL)

# Expires on a given date and time
file_id <- write_civis_file(iris, expires_at = "2030-01-01")
file_id <- write_civis_file(iris, expires_at = "12:00:00")
file_id <- write_civis_file(iris, expires_at = "2030-01-01 12:00:00")

# Upload a large file in parallel.
library(future)
plan(multisession)
file_id <- write_civis_file("my_large_file")
# }